% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_decode.R
\name{apply_deltas_to_anchors}
\alias{apply_deltas_to_anchors}
\title{Apply box deltas to anchors}
\usage{
apply_deltas_to_anchors(anchors, deltas)
}
\arguments{
\item{anchors}{Tensor of shape (N, 4), in \link{x1, y1, x2, y2} format}

\item{deltas}{Tensor of shape (N, 4), each row is \link{dx, dy, dw, dh}}
}
\value{
Tensor of shape (N, 4) with predicted boxes in \link{x1, y1, x2, y2} format
}
\description{
Applies \link{dx, dy, dw, dh} deltas to anchor boxes to get predicted boxes.
}
