% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder-dataset.R
\name{base_loader}
\alias{base_loader}
\title{Base loader}
\usage{
base_loader(path)
}
\arguments{
\item{path}{path to the image to load from}
}
\description{
Loads an image using \code{jpeg}, or \code{png} packages depending on the
file extension.
}
