% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-boxes.R
\name{box_area}
\alias{box_area}
\title{Box Area}
\usage{
box_area(boxes)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes for which the area will be computed. They
are expected to be in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format with
\itemize{
\item \eqn{0 \leq x_{min} < x_{max}} and
\item \eqn{0 \leq y_{min} < y_{max}}.
}}
}
\value{
area (Tensor[N]): area for each box
}
\description{
Computes the area of a set of bounding boxes, which are specified by its
\eqn{(x_{min}, y_{min}, x_{max}, y_{max})} coordinates.
}
