% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_cxcywh_to_xyxy}
\alias{box_cxcywh_to_xyxy}
\title{box_cxcywh_to_xyxy}
\usage{
box_cxcywh_to_xyxy(boxes)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in \eqn{(c_x, c_y, w, h)} format which will be converted.}
}
\value{
boxes (Tensor(N, 4)): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format.
}
\description{
Converts bounding boxes from  \eqn{(c_x, c_y, w, h)} format to \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format.
\eqn{(c_x, c_y)} refers to center of bounding box
(w, h) are width and height of bounding box
}
