% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_xywh_to_xyxy}
\alias{box_xywh_to_xyxy}
\title{box_xywh_to_xyxy}
\usage{
box_xywh_to_xyxy(boxes)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in (x, y, w, h) which will be converted.}
}
\value{
boxes (Tensor[N, 4]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format.
}
\description{
Converts bounding boxes from  (x, y, w, h) format to \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format.
(x, y) refers to top left of bouding box.
(w, h) refers to width and height of box.
}
