% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_xyxy_to_cxcywh}
\alias{box_xyxy_to_cxcywh}
\title{box_xyxy_to_cxcywh}
\usage{
box_xyxy_to_cxcywh(boxes)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format which will be converted.}
}
\value{
boxes (Tensor(N, 4)): boxes in \eqn{(c_x, c_y, w, h)} format.
}
\description{
Converts bounding boxes from  \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format to \eqn{(c_x, c_y, w, h)} format.
(x1, y1) refer to top left of bounding box
(x2, y2) refer to bottom right of bounding box
}
