% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_xyxy_to_xywh}
\alias{box_xyxy_to_xywh}
\title{box_xyxy_to_xywh}
\usage{
box_xyxy_to_xywh(boxes)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in (x1, y1, x2, y2) which will be converted.}
}
\value{
boxes (Tensor[N, 4]): boxes in (x, y, w, h) format.
}
\description{
Converts bounding boxes from  (x1, y1, x2, y2) format to (x, y, w, h) format.
(x1, y1) refer to top left of bounding box
(x2, y2) refer to bottom right of bounding box
}
