% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_xyxy_to_xyxyr}
\alias{box_xyxy_to_xyxyr}
\title{box_xyxy_to_xyxyr}
\usage{
box_xyxy_to_xyxyr(boxes, angle = 0)
}
\arguments{
\item{boxes}{(Tensor[N, 4] or [N, 5]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})}
or \eqn{(x_{min}, y_{min}, x_{max}, y_{max}, r_{deg})} format, where \eqn{r_{deg}} format.}

\item{angle}{(numeric): Rotation angle in degrees (anti-clockwise).
A single numeric value or a tensor of shape (N,). Default is 0.}
}
\value{
(Tensor[N, 5]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max}, r_{deg})} format with accumulated angle.
}
\description{
Converts bounding boxes from \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format to
\eqn{(x_{min}, y_{min}, x_{max}, y_{max}, r_{deg})} format, where \eqn{r_{deg}} is the rotation
angle in degrees (anti-clockwise). Handles composition natively.
}
