% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-rf-detr.R
\name{calculate_uncertainty}
\alias{calculate_uncertainty}
\title{Compute Uncertainty}
\usage{
calculate_uncertainty(logits)
}
\arguments{
\item{logits}{(Tensor): A tensor of shape (R, 1, ...) for class-specific or
class-agnostic, where R is the total number of predicted masks in all images and C is
the number of foreground classes. The values are logits.}
}
\value{
A tensor of shape (R, 1, ...) that contains uncertainty scores with
the most uncertain locations having the highest uncertainty score.
}
\description{
We estimate uncertainty as L1 distance between 0.0 and the logit prediction in 'logits' for the
foreground class in \code{classes}.
}
