% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-caltech.R
\name{caltech101_dataset}
\alias{caltech101_dataset}
\title{Caltech101 Dataset}
\usage{
caltech101_dataset(
  root,
  target_type = "category",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{(string): Root directory where the dataset will be stored or exists.}

\item{target_type}{(string or list): Type of target to return. Can be "category", "annotation",
or a list of both. "category" returns the class label, "annotation" returns contour points.}

\item{transform}{(callable, optional): A function/transform to apply to the image.}

\item{target_transform}{(callable, optional): A function/transform to apply to the target(s).}

\item{download}{(bool): If TRUE, downloads the dataset from the internet.}
}
\description{
\href{https://data.caltech.edu/records/20086}{Caltech101} Dataset.
}
