% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-caltech.R
\name{caltech256_dataset}
\alias{caltech256_dataset}
\title{Caltech 256 Dataset}
\usage{
caltech256_dataset(
  root = rappdirs::user_cache_dir("torch"),
  transform = NULL,
  target_transform = NULL,
  download = FALSE,
  image_size = c(224, 224)
)
}
\arguments{
\item{root}{Character. Root directory for dataset storage. The dataset will be stored under \code{root/caltech101}.}

\item{transform}{Optional function to transform input images after loading.}

\item{target_transform}{Optional function to transform labels.}

\item{download}{Logical. Whether to download the dataset if not found locally. Default is \code{FALSE}.}

\item{image_size}{Integer vector of length 2. Size to resize images to (width, height). Default is \code{c(224, 224)}.}
}
\value{
A caltech256_dataset object representing the dataset.
}
\description{
Caltech 256 Dataset
}
\details{
Loads the Caltech 256 Dataset, containing images from 256 object categories.
The dataset includes images and their corresponding class labels.
There is no built-in train/test split.
}
\examples{
\dontrun{
root_dir <- tempfile()
caltech <- caltech256_dataset(root = root_dir, download = TRUE)
first_item <- caltech[1]
# image tensor of first item
first_item$x
# label (class name) of first item
first_item$y
}

}
