% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-caltech.R
\name{caltech_classes}
\alias{caltech_classes}
\title{Caltech Class Labels}
\usage{
caltech_classes(class_id = 1:257)
}
\arguments{
\item{class_id}{Integer vector of 1-based class identifiers.}
}
\value{
A character vector with 257 entries representing the Caltech 257
class labels.
}
\description{
Utilities for resolving Caltech class identifiers to their corresponding
human readable labels.
}
\seealso{
Other class_resolution: 
\code{\link{coco_classes}()},
\code{\link{imagenet_classes}()},
\code{\link{pascal_voc_classes}()}
}
\concept{class_resolution}
