% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-cifar.R
\name{cifar10_dataset}
\alias{cifar10_dataset}
\alias{cifar100_dataset}
\title{CIFAR datasets}
\usage{
cifar10_dataset(
  root,
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)

cifar100_dataset(
  root,
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{(string): Root directory of dataset where directory
\code{cifar-10-batches-bin} exists or will be saved to if download is set to TRUE.}

\item{train}{(bool, optional): If True, creates dataset from
\code{training.pt}, otherwise from \code{test.pt}.}

\item{transform}{(callable, optional): A function/transform that  takes in an
PIL image and returns a transformed version. E.g,
\code{\link[=transform_random_crop]{transform_random_crop()}}.}

\item{target_transform}{(callable, optional): A function/transform that takes
in the target and transforms it.}

\item{download}{(bool, optional): If true, downloads the dataset from the
internet and puts it in root directory. If dataset is already downloaded,
it is not downloaded again.}
}
\value{
A torch::dataset object. Each item is a list with:
\itemize{
\item \code{x}: a 32x32x3 integer array
\item \code{y}: the class label
}
}
\description{
Downloads and prepares the
\href{https://www.cs.toronto.edu/~kriz/cifar.html}{CIFAR10} dataset.

Downloads and prepares the
\href{https://www.cs.toronto.edu/~kriz/cifar.html}{CIFAR100} dataset.
}
\examples{
\dontrun{
ds <- cifar10_dataset(root = tempdir(), download = TRUE)
item <- ds[1]
item$x
item$y
}
}
\seealso{
Other datasets: 
\code{\link{caltech_dataset}},
\code{\link{coco_caption_dataset}()},
\code{\link{coco_detection_dataset}()},
\code{\link{eurosat_datasets}},
\code{\link{fer_dataset}()},
\code{\link{fgvc_aircraft_dataset}()},
\code{\link{flowers102_dataset}()},
\code{\link{image_folder_dataset}()},
\code{\link{tiny_imagenet_dataset}()}
}
\concept{datasets}
