% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-boxes.R
\name{clip_boxes_to_image}
\alias{clip_boxes_to_image}
\title{Clip Boxes to Image}
\usage{
clip_boxes_to_image(boxes, size)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in \verb{(x1, y1, x2, y2)} format
with \verb{0 <= x1 < x2} and \verb{0 <= y1 < y2}.}

\item{size}{(Tuple[height, width]): size of the image}
}
\value{
clipped_boxes (Tensor[N, 4])
}
\description{
Clip boxes so that they lie inside an image of size \code{size}.
}
