% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-coco.R
\name{coco_caption_dataset}
\alias{coco_caption_dataset}
\title{COCO Caption Dataset}
\usage{
coco_caption_dataset(root, train = TRUE, year = c("2014"), download = FALSE)
}
\arguments{
\item{root}{Root directory where the dataset is stored or will be downloaded to.}

\item{train}{(bool, optional): If True, creates dataset from
\code{training.pt}, otherwise from \code{test.pt}.}

\item{year}{Character. Dataset version year. One of \code{"2014"}, \code{"2016"}, or \code{"2017"}.}

\item{download}{(bool, optional): If true, downloads the dataset from the
internet and puts it in root directory. If dataset is already downloaded,
it is not downloaded again.}
}
\value{
An object of class \code{coco_caption_dataset}. Each item is a list:
\itemize{
\item \code{x}: an \verb{(H, W, C)} numeric array containing the RGB image.
\item \code{y}: a character string with the image caption.
}
}
\description{
Loads the MS COCO dataset for image captioning.
}
\examples{
\dontrun{
ds <- coco_caption_dataset(
  root = "~/data",
  train = FALSE,
  download = TRUE
)
example <- ds[1]

# Access image and caption
x <- example$x
y <- example$y

# Prepare image for plotting
image_array <- as.numeric(x)
dim(image_array) <- dim(x)

plot(as.raster(image_array))
title(main = y, col.main = "black")
}
}
\seealso{
Other datasets: 
\code{\link{caltech_dataset}},
\code{\link{cifar10_dataset}()},
\code{\link{coco_detection_dataset}()},
\code{\link{eurosat_datasets}},
\code{\link{fer_dataset}()},
\code{\link{fgvc_aircraft_dataset}()},
\code{\link{flowers102_dataset}()},
\code{\link{image_folder_dataset}()},
\code{\link{tiny_imagenet_dataset}()}
}
\concept{datasets}
