% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-coco.R
\name{coco_classes}
\alias{coco_classes}
\title{MS COCO Class Labels}
\usage{
coco_classes(class_id = 1:90)
}
\arguments{
\item{class_id}{Integer vector of 1-based class identifiers.}
}
\value{
A character vector with the COCO class names
}
\description{
Utilities for resolving COCO 90 class identifiers to their corresponding
human readable labels. The labels are retrieved from pytorch/vision source to be compliant
with torchvision pretrained models.
}
\seealso{
Other class_resolution: 
\code{\link{caltech_classes}()},
\code{\link{imagenet_classes}()},
\code{\link{pascal_voc_classes}()}
}
\concept{class_resolution}
