% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco-dataset.R
\name{coco_dataset}
\alias{coco_dataset}
\title{MS COCO Dataset}
\usage{
coco_dataset(root, split = "train", year = "2017", download = FALSE, ...)
}
\arguments{
\item{root}{Root directory where data is downloaded or stored.}

\item{split}{One of "train", "val".}

\item{year}{Dataset year, e.g., "2017".}

\item{download}{Whether to download the dataset.}

\item{...}{Additional arguments passed to the dataset.}
}
\description{
Loads MS COCO Detection Dataset for object detection and segmentation.
}
