% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_model.R
\name{create_model}
\alias{create_model}
\title{Create a neural network model}
\usage{
create_model(input_dim = 57, hidden_dim = 64, output_dim = 2)
}
\arguments{
\item{input_dim}{Number of input features}

\item{hidden_dim}{Hidden layer size}

\item{output_dim}{Number of output classes}
}
\value{
A torch model
}
\description{
Create a neural network model
}
