% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-coco-detection.R
\name{dataset_coco_detection}
\alias{dataset_coco_detection}
\title{COCO Detection Dataset}
\usage{
dataset_coco_detection(
  root,
  train = TRUE,
  year = c("2017", "2016", "2014"),
  download = FALSE,
  transforms = NULL,
  target_transform = NULL
)
}
\arguments{
\item{root}{Root directory for data.}

\item{train}{Whether to load training split (TRUE) or validation (FALSE).}

\item{year}{Dataset year: "2014", "2016", or "2017".}

\item{download}{If TRUE, downloads the dataset if needed.}

\item{transforms}{Transform applied to image.}

\item{target_transform}{Transform applied to target.}
}
\description{
Loads MS COCO Dataset for object detection and segmentation.
}
