% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-mobilenetv3.R
\name{.rename_mobilenet_v3_state_dict}
\alias{.rename_mobilenet_v3_state_dict}
\title{Rename state_dict keys to match current MobileNetV3 model structure}
\usage{
.rename_mobilenet_v3_state_dict(state_dict)
}
\arguments{
\item{state_dict}{Named list of model parameters}
}
\value{
Named list with updated key names
}
\description{
Rename state_dict keys to match current MobileNetV3 model structure
}
\keyword{internal}
