% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-eurosat.R
\name{eurosat_datasets}
\alias{eurosat_datasets}
\alias{eurosat_dataset}
\alias{eurosat_all_bands_dataset}
\alias{eurosat100_dataset}
\title{EuroSAT Dataset}
\usage{
eurosat_dataset(
  root,
  split = "train",
  download = FALSE,
  transform = NULL,
  target_transform = NULL
)

eurosat_all_bands_dataset(
  root,
  split = "train",
  download = FALSE,
  transform = NULL,
  target_transform = NULL
)

eurosat100_dataset(
  root,
  split = "train",
  download = FALSE,
  transform = NULL,
  target_transform = NULL
)
}
\arguments{
\item{root}{(Optional) Character. The root directory where the dataset will be stored.
if empty, will use the default \code{rappdirs::user_cache_dir("torch")}.}

\item{split}{Character. Must be one of \code{train}, \code{val}, or \code{test}.}

\item{download}{(bool, optional): If true, downloads the dataset from the
internet and puts it in root directory. If dataset is already downloaded,
it is not downloaded again.}

\item{transform}{(callable, optional): A function/transform that  takes in an
PIL image and returns a transformed version. E.g,
\code{\link[=transform_random_crop]{transform_random_crop()}}.}

\item{target_transform}{(callable, optional): A function/transform that takes
in the target and transforms it.}
}
\value{
A \code{torch::dataset} object. Each item is a list with:
\itemize{
\item \code{x}: a 64x64 image tensor with 3 (RGB) or 13 (all bands) channels
\item \code{y}: the class label
}
}
\description{
Downloads and prepare the EuroSAT dataset from Hugging Face datasets.
The dataset consists of Land Use and Land Cover Classification with Sentinel-2
satellite images. Images are openly and freely made available by the Earth
observation program Copernicus. Images are organized into 10 classes.

Downloads and prepares the EuroSAT dataset with 13 spectral bands.

A subset of 100 images with 13 spectral bands useful for workshops and demos.
}
\details{
\code{eurosat_dataset()} provides a total of 27,000 RGB labeled images.

\code{eurosat_all_bands_dataset()} provides a total of 27,000 labeled images with 13 spectral channel bands.

\code{eurosat100_dataset()} provides a subset of 100 labeled images with 13 spectral channel bands.
}
\examples{
\dontrun{
# Initialize the dataset
ds <- eurosat100_dataset(split = "train", download = TRUE)

# Access the first item
head <- ds[1]
print(head$x) # Image
print(head$y) # Label
}
}
\seealso{
Other datasets: 
\code{\link{caltech_dataset}},
\code{\link{cifar10_dataset}()},
\code{\link{coco_caption_dataset}()},
\code{\link{coco_detection_dataset}()},
\code{\link{fer_dataset}()},
\code{\link{fgvc_aircraft_dataset}()},
\code{\link{flowers102_dataset}()},
\code{\link{image_folder_dataset}()},
\code{\link{tiny_imagenet_dataset}()}
}
\concept{datasets}
