% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate model on test data}
\usage{
evaluate_model(model, test_dl)
}
\arguments{
\item{model}{The trained model}

\item{test_dl}{Dataloader for test set}
}
\value{
List with confusion matrix and accuracy
}
\description{
Evaluate model on test data
}
