% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-mnist.R
\name{fashion_mnist_dataset}
\alias{fashion_mnist_dataset}
\title{Fashion-MNIST dataset}
\usage{
fashion_mnist_dataset(
  root,
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{(string): Root directory of dataset where
\code{FashionMNIST/processed/training.pt} and \code{FashionMNIST/processed/test.pt} exist.}

\item{train}{(bool, optional): If TRUE, creates dataset from \code{training.pt},
otherwise from \code{test.pt}.}

\item{transform}{(callable, optional): A function/transform that takes in an
image and returns a transformed version. E.g., \code{\link[=transform_random_crop]{transform_random_crop()}}.}

\item{target_transform}{(callable, optional): A function/transform that takes
in the target and transforms it.}

\item{download}{(bool, optional): If TRUE, downloads the dataset from the
internet and puts it in root directory. If dataset is already downloaded,
it is not downloaded again.}
}
\description{
Prepares the \href{https://github.com/zalandoresearch/fashion-mnist}{Fashion-MNIST} dataset
and optionally downloads it.
}
\details{
Fashion-MNIST dataset
}
\seealso{
\code{\link[=mnist_dataset]{mnist_dataset()}}, \code{\link[=kmnist_dataset]{kmnist_dataset()}}
}
