% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-fer.R
\name{fer_dataset}
\alias{fer_dataset}
\title{FER2013 dataset}
\usage{
fer_dataset(
  root = rappdirs::user_cache_dir("torch"),
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Character. Root directory for dataset storage (default folder: \verb{root/fer2013/processed/}).}

\item{train}{Logical. If \code{TRUE}, loads the training split; otherwise, loads the test split. Default is \code{TRUE}.}

\item{transform}{Optional function to transform input image tensors.}

\item{target_transform}{Optional function to transform target labels.}

\item{download}{Logical. Whether to download the dataset if it is not found locally. Default is \code{FALSE}.}
}
\value{
A FER2013 dataset object.
}
\description{
FER2013 Dataset
}
\details{
Loads the FER2013 dataset, a facial expression recognition dataset containing
grayscale 48x48 pixel images labeled with one of 7 emotions.

The dataset consists of two splits:
\itemize{
\item "train": training set
\item "test": test set
}
}
\examples{
\dontrun{
fer <- fer_dataset(train = TRUE, download = TRUE)
first_item <- fer[1]
# image tensor of the first sample
first_item$x
# label tensor of the first sample
first_item$y
}

}
