% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-fgvc.R
\name{fgvc_aircraft_dataset}
\alias{fgvc_aircraft_dataset}
\title{FGVC Aircraft dataset}
\usage{
fgvc_aircraft_dataset(
  root = rappdirs::user_cache_dir("torch"),
  split = "train",
  annotation_level = "variant",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Character. Root directory for dataset storage. The dataset will be stored under \verb{root/fgvc-aircraft-2013b}.}

\item{split}{Character. Dataset split to use. One of \code{"train"}, \code{"trainval"}, or \code{"test"}. Default is \code{"train"}.}

\item{annotation_level}{Character. Level of annotation to use. One of \code{"variant"}, \code{"family"}, or \code{"manufacturer"}. Default is \code{"variant"}.}

\item{transform}{Optional function to transform input images after loading.}

\item{target_transform}{Optional function to transform labels.}

\item{download}{Logical. Whether to download the dataset if not found locally. Default is \code{FALSE}.}
}
\value{
A fgvc_aircraft_dataset object representing the dataset.
}
\description{
fgvc Aircraft Dataset
}
\details{
Loads the fgvc Aircraft Dataset, a dataset of aircraft images with multiple annotation levels:
\itemize{
\item "variant": fine-grained aircraft variants (e.g., specific model variants)
\item "family": aircraft families
\item "manufacturer": aircraft manufacturers
}

The dataset supports different splits: "train", "trainval", and "test".
}
\examples{
\dontrun{
root_dir <- tempfile()
fgvc <- fgvc_aircraft_dataset(root = root_dir, split = "train", annotation_level = "variant", download = TRUE)
first_item <- fgvc[1]
# image tensor of first item
first_item$x
# label of first item
first_item$y
#class name of the first item
first_item$class_name
}

}
