% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-flickr.R
\name{flickr30k_dataset}
\alias{flickr30k_dataset}
\title{Flickr30k Dataset}
\usage{
flickr30k_dataset(
  root = tempdir(),
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Character. Root directory where the dataset will be stored under \code{root/flickr30k}.}

\item{train}{Logical. If \code{TRUE}, loads the training split, otherwise the test split. Default is \code{TRUE}.}

\item{transform}{Optional function to transform input images after loading.}

\item{target_transform}{Optional function to transform captions.}

\item{download}{Logical. Whether to download the dataset if not found locally. Default is \code{FALSE}.}
}
\value{
A flickr30k_dataset object representing the dataset with images and their corresponding captions.
}
\description{
Flickr30k Dataset
}
\details{
Loads the Flickr30k dataset consisting of 31,014 images, each annotated with five captions.
The dataset is split into training and test sets using the official Karpathy splits.
Images are resized to a fixed 224x224 resolution, and captions are extracted as lists of raw strings.
}
\examples{
\dontrun{
root_dir <- tempfile()
flickr30k <- flickr30k_dataset(download = TRUE)
first_item <- flickr30k[1]
# Tensor representing the image
first_item$x
# List of captions for the image
first_item$y
}

}
