% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transforms-geometry.R
\name{item_transform_perspective}
\alias{item_transform_perspective}
\title{Apply a perspective transform to a dataset item}
\usage{
item_transform_perspective(
  x,
  startpoints,
  endpoints,
  interpolation = 2,
  fill = NULL
)
}
\arguments{
\item{x}{A dataset item, typically an \code{image_with_bounding_box} or
\code{image_with_segmentation_mask} object containing an image tensor
and associated target data.}

\item{startpoints}{(list of integer vectors of length 2): List containing
four integer vectors of length 2 corresponding to four corners
\verb{[top-left, top-right, bottom-right, bottom-left]} of the original image.}

\item{endpoints}{(list of integer vectors of length 2): List containing
four integer vectors of length 2 corresponding to four corners
\verb{[top-left, top-right, bottom-right, bottom-left]} of the transformed
image.}

\item{interpolation}{(integer, optional): Desired interpolation. An integer
\code{0 = nearest}, \code{2 = bilinear}, and \code{3 = bicubic} or a name from
\code{\link[magick:filter_types]{magick::filter_types()}}.}

\item{fill}{(numeric vector or numeric): Pixel fill value for constant fill.
Default is \code{0}. If a numeric vector of length 3, it is used to fill R, G,
B channels respectively. This value is only used when the \code{padding_mode}
is constant. Only integer values are supported for Tensors.}
}
\value{
A dataset item of the same class with the image and target
perspective transformed.
}
\description{
Applies a perspective transform to the image inside a dataset item, as
defined by \code{\link{transform_perspective}}. For detection items,
bounding box coordinates are transformed accordingly and clamped to the
image canvas, boxes that end up completely outside the image are removed.
For segmentation items, both the image and the masks are transformed,
using nearest neighbour interpolation for the masks.
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/b/b6/Felis_catus-cat_on_snow.jpg"
img <- base_loader(url) |> transform_to_tensor()

boxes <- torch_tensor(matrix(c(600, 200, 2880, 1860), ncol = 4), dtype = torch_float32())

before <- list(x = img, y = list(boxes = boxes, labels = "cat"))
class(before) <- c("image_with_bounding_box", "list")

startpoints <- list(c(0, 0), c(2880, 0), c(2880, 1860), c(0, 1860))
endpoints <- list(c(100, 50), c(2700, 0), c(2800, 1800), c(80, 1900))

after <- item_transform_perspective(before, startpoints, endpoints)

before_plot <- draw_bounding_boxes(before, colors = "blue", width = 10)$to(torch_float())$div(255)
after_plot <- draw_bounding_boxes(after, colors = "red", width = 10)$to(torch_float())$div(255)

grid <- vision_make_grid(torch_stack(list(before_plot, after_plot)), scale = TRUE)
tensor_image_browse(grid)
}

}
\seealso{
Other item_unitary_transforms:
\code{\link[=item_transform_affine]{item_transform_affine()}},
\code{\link[=item_transform_center_crop]{item_transform_center_crop()}},
\code{\link[=item_transform_crop]{item_transform_crop()}},
\code{\link[=item_transform_hflip]{item_transform_hflip()}},
\code{\link[=item_transform_pad]{item_transform_pad()}},
\code{\link[=item_transform_resize]{item_transform_resize()}},
\code{\link[=item_transform_rotate]{item_transform_rotate()}},
\code{\link[=item_transform_vflip]{item_transform_vflip()}}
}
\concept{item_unitary_transforms}
