% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transforms-geometry.R
\name{item_transform_resize}
\alias{item_transform_resize}
\title{Resize a dataset item}
\usage{
item_transform_resize(x, size, interpolation = 2)
}
\arguments{
\item{x}{A dataset item, typically an \code{image_with_bounding_box} or
\code{image_with_segmentation_mask} object containing an image tensor
and associated target data.}

\item{size}{(integer vector or integer): Desired output size. If \code{size} is
an integer vector of length 2 like \code{c(h, w)}, output size will be matched
to this. If \code{size} is a bare integer, smaller edge of the image will be
matched to this number, i.e., if height > width, then image will be
rescaled to \verb{(size * height / width, size)}.}

\item{interpolation}{(integer, optional): Desired interpolation. An integer
\code{0 = nearest}, \code{2 = bilinear}, and \code{3 = bicubic} or a name from
\code{\link[magick:filter_types]{magick::filter_types()}}.}
}
\value{
A dataset item of the same class with the image and target resized.
}
\description{
Resizes the image inside a dataset item. For detection items, bounding box
coordinates are rescaled by the same factors as the image. For segmentation
items, the masks are resized alongside the image with nearest-neighbour
sampling, so that they keep their discrete values.
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/b/b6/Felis_catus-cat_on_snow.jpg"
img <- base_loader(url) |> transform_to_tensor()

boxes <- torch_tensor(matrix(c(600, 200, 2880, 1860), ncol = 4), dtype = torch_float32())

before <- list(x = img, y = list(boxes = boxes, labels = "cat"))
class(before) <- c("image_with_bounding_box", "list")

after <- item_transform_resize(before, size = c(600, 800))

before_plot <- draw_bounding_boxes(before, colors = "blue", width = 10)
after_plot <- draw_bounding_boxes(after, colors = "red", width = 10)
tensor_image_browse(before_plot)
tensor_image_browse(after_plot)
}

}
\seealso{
Other item_unitary_transforms:
\code{\link[=item_transform_affine]{item_transform_affine()}},
\code{\link[=item_transform_center_crop]{item_transform_center_crop()}},
\code{\link[=item_transform_crop]{item_transform_crop()}},
\code{\link[=item_transform_hflip]{item_transform_hflip()}},
\code{\link[=item_transform_pad]{item_transform_pad()}},
\code{\link[=item_transform_perspective]{item_transform_perspective()}},
\code{\link[=item_transform_rotate]{item_transform_rotate()}},
\code{\link[=item_transform_vflip]{item_transform_vflip()}}
}
\concept{item_unitary_transforms}
