% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-rf-detr.R
\name{lwdetr}
\alias{lwdetr}
\title{The forward expects a NestedTensor, which consists of:
- samples$tensor: batched images, of shape [batch_size x 3 x H x W]
- samples$mask: a binary mask of shape [batch_size x H x W], containing 1 on padded pixels}
\usage{
lwdetr(
  backbone,
  transformer,
  segmentation_head,
  num_classes,
  num_queries,
  aux_loss = FALSE,
  group_detr = 1,
  two_stage = FALSE,
  lite_refpoint_refine = FALSE,
  bbox_reparam = FALSE
)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        It returns a dict with the following elements:
           - "pred_logits": the classification logits (including no-object) for all queries.
                            Shap = \[batch_size x num_queries x num_classes]
           - "pred_boxes": The normalized boxes coordinates for all queries, represented as
                           (center_x, center_y, width, height). These values are normalized in [0, 1],
                           relative to the size of each individual image (disregarding possible padding).
                           See PostProcess for information on how to retrieve the unnormalized bounding box.
           - "aux_outputs": Optional, only returned when auxilary losses are activated. It is a list of
                            dictionnaries containing the two above keys for each decoder layer.
}\if{html}{\out{</div>}}
}
