% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-mobilenetv3_large.R
\name{model_mobilenet}
\alias{model_mobilenet}
\alias{model_mobilenet_v3_large_quantized}
\title{Quantized MobileNetV3 Large model architecture.}
\usage{
model_mobilenet_v3_large_quantized(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{(bool): If TRUE, returns a model pre-trained on ImageNet.}

\item{progress}{(bool): If TRUE, displays a progress bar of the download to
stderr.}

\item{...}{Other parameters passed to the model implementation.}
}
\description{
This function mirrors \verb{torchvision::quantization::mobilenet_v3_large} and
loads quantized weights when \code{pretrained} is \code{TRUE}.
}
\seealso{
Other classification_model: 
\code{\link{model_alexnet}()},
\code{\link{model_convnext}},
\code{\link{model_efficientnet}},
\code{\link{model_efficientnet_v2}},
\code{\link{model_facenet}},
\code{\link{model_inception_v3}()},
\code{\link{model_maxvit}()},
\code{\link{model_mobilenet_v2}()},
\code{\link{model_mobilenet_v3}},
\code{\link{model_resnet}},
\code{\link{model_vgg}},
\code{\link{model_vit}}
}
\concept{classification_model}
