% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-mobilenetv3_large.R
\name{model_mobilenet_v3_large}
\alias{model_mobilenet_v3_large}
\title{Constructs a MobileNetV3 Large architecture.}
\usage{
model_mobilenet_v3_large(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{(bool): If TRUE, returns a model pre-trained on ImageNet.}

\item{progress}{(bool): If TRUE, displays a progress bar of the download to
stderr.}

\item{...}{Other parameters passed to the model implementation.}
}
\description{
This implementation mimics \code{torchvision::mobilenet_v3_large}.
}
\concept{models}
