% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-convnext.R
\name{model_resnext}
\alias{model_resnext}
\title{ConvNeXt Implementation}
\arguments{
\item{...}{Other parameters passed to the model implementation.
\subsection{Model Summary and Performance for pretrained weights}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{| Model                | Top-1 Acc| Params  | GFLOPS | File Size | `num_classes` | image size |
|----------------------|----------|---------|-------|-----------|----------------|------------|
| convnext_tiny_1k     | 82.1\%    | 28M    | 4.5    | 21.1 MB   |  1000 | 224 x 224 |
| convnext_tiny_22k    | 82.9\%    | 29M    | 4.5    | 9.8 MB    | 21841 | 224 x 224 |
| convnext_small_22k   | 84.6\%    | 50M    | 8.7    | 9.8 MB    | 21841 | 224 x 224 |
| convnext_small_22k1k | 84.6\%    | 50M    | 8.7    | 9.8 MB    | 1000  | 224 x 224 |
| convnext_base_1k     | 85.1\%    | 89M    | 15.4   | 9.8 MB    | 1000  | 224 x 224 |
| convnext_base_22k    | 85.8\%    | 89M    | 15.4   | 9.8 MB    | 21841 | 224 x 224 |
| convnext_large_1k    | 84.3\%    | 198M   | 34.4   | 9.8 MB    | 1000  | 224 x 224 |
| convnext_large_22k   | 86.6\%    | 198M   | 34.4   | 9.8 MB    | 21841 | 224 x 224 |
}\if{html}{\out{</div>}}
}}
}
\description{
Implements the ConvNeXt architecture from \href{https://arxiv.org/pdf/2201.03545}{ConvNeXt: A ConvNet for the 2020s}
}
\seealso{
Other classification_model: 
\code{\link{model_alexnet}()},
\code{\link{model_efficientnet}},
\code{\link{model_efficientnet_v2}},
\code{\link{model_inception_v3}()},
\code{\link{model_maxvit}()},
\code{\link{model_mobilenet_v2}()},
\code{\link{model_mobilenet_v3}},
\code{\link{model_resnet}},
\code{\link{model_vgg}},
\code{\link{model_vit}}
}
\concept{classification_model}
