% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-coco-caption.R
\name{ms_coco_captions_dataset}
\alias{ms_coco_captions_dataset}
\title{MS COCO Captions Dataset}
\usage{
ms_coco_captions_dataset(
  root,
  train = TRUE,
  year = c("2014"),
  download = FALSE
)
}
\arguments{
\item{root}{Root directory for the dataset.}

\item{train}{Whether to load training data (TRUE) or validation data (FALSE).}

\item{year}{One of "2014" (only available year for captions).}

\item{download}{If TRUE, downloads the dataset if needed.}
}
\description{
Loads the Microsoft COCO dataset for image captioning.
}
