% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\docType{data}
\name{ops}
\alias{ops}
\title{Torchvision ops}
\format{
An object of class \code{list} of length 4.
}
\usage{
ops
}
\description{
Provides access to common bounding box conversion utilities.
}
\keyword{datasets}
