% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-rf-detr.R
\name{post_process}
\alias{post_process}
\title{PostProcess to coco API}
\usage{
post_process(num_select = 300)
}
\arguments{
\item{outputs:}{raw outputs of the model}

\item{target_sizes:}{tensor of dimension \link{batch_size x 2} containing the size of each images of the batch
For evaluation, this must be the original image size (before any data augmentation)
For visualization, this should be the image size after data augment, but before padding}
}
\description{
converts the model's output into the format expected by the coco api
}
