% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_model.R
\name{prepare_datasets}
\alias{prepare_datasets}
\title{Prepare train/test datasets}
\usage{
prepare_datasets(data, train_ratio = 0.8)
}
\arguments{
\item{data}{List returned from load_spam_data}

\item{train_ratio}{Fraction of data to use for training}
}
\value{
A list with train/test datasets and scaler info
}
\description{
Prepare train/test datasets
}
