% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-boxes.R
\name{remove_small_boxes}
\alias{remove_small_boxes}
\title{Remove Small Boxes}
\usage{
remove_small_boxes(boxes, min_size)
}
\arguments{
\item{boxes}{(Tensor\link{N, 4}): boxes in \verb{(x1, y1, x2, y2)} format
with \verb{0 <= x1 < x2} and \verb{0 <= y1 < y2}.}

\item{min_size}{(float): minimum size}
}
\value{
keep (Tensor\link{K}): indices of the boxes that have both sides
larger than min_size
}
\description{
Remove boxes which contains at least one side smaller than min_size.
}
