% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-rf-detr.R
\name{set_criterion}
\alias{set_criterion}
\title{Create the criterion}
\usage{
set_criterion(
  num_classes,
  matcher,
  weight_dict,
  focal_alpha,
  losses,
  group_detr = 1,
  sum_group_losses = FALSE,
  use_varifocal_loss = FALSE,
  use_position_supervised_loss = FALSE,
  ia_bce_loss = FALSE,
  mask_point_sample_ratio = 16L
)
}
\arguments{
\item{num_classes}{number of object categories, omitting the special no-object category}

\item{matcher}{module able to compute a matching between targets and proposals}

\item{weight_dict}{dict containing as key the names of the losses and as values their relative weight.}

\item{focal_alpha}{alpha in Focal Loss}

\item{losses}{list of all the losses to be applied. See get_loss for list of available losses.}

\item{group_detr}{Number of groups to speed detr training. Default is 1.}

\item{mask_point_sample_ratio}{value of the mask-point sample ratio
Classification loss (Binary focal loss)}

\item{sum_group_losses:}{whether or not to sum the losses}

\item{use_varifocal_loss:}{whether or not to use varifocal loss}

\item{use_position_supervised_loss:}{whether or not to use positional supervsed loss}

\item{ia_bce_loss:}{whether or not to use ia binary cros entropy loss}

\item{targets}{a named list that must include a "labels" named tensor of dim [nb_target_boxes]
Compute the cardinality error, i.e. the absolute error in the number of predicted non-empty boxes
This is not really a loss, it is intended for logging purposes only. It doesn't propagate gradients
Compute the losses related to the bounding boxes, the L1 regression loss and the GIoU loss
targets dicts must contain the key "boxes" containing a tensor of dim \link{nb_target_boxes, 4}
The target boxes are expected in format (center_x, center_y, w, h), normalized by the image size.
Compute BCE-with-logits and Dice losses for segmentation masks on matched pairs.
Expects outputs to contain 'pred_masks' of shape c(B, Q, H, W) and targets with key 'masks'.
performs the loss computation.}

\item{outputs:}{list of tensors, see the output specification of the model for the format}

\item{targets:}{list of lists, such that length(targets) == batch_size.
The expected names in each list depends on the losses applied, see each loss' doc}
}
\description{
Create the criterion
}
