% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-rf-detr.R
\name{sigmoid_ce_loss}
\alias{sigmoid_ce_loss}
\title{Sigmoid cross-entropy loss}
\usage{
sigmoid_ce_loss(inputs, targets, num_masks)
}
\arguments{
\item{inputs:}{A float tensor of arbitrary shape.
The predictions for each example.}

\item{targets:}{A float tensor with the same shape as inputs. Stores the binary
classification label for each element in inputs
(0 for the negative class and 1 for the positive class).}

\item{num_masks:}{number of masks}
}
\value{
Loss tensor
}
\description{
Sigmoid cross-entropy loss
}
