% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision-utils.R
\name{tensor_image_browse}
\alias{tensor_image_browse}
\title{Display image tensor}
\usage{
tensor_image_browse(image, browser = getOption("browser"))
}
\arguments{
\item{image}{\code{torch_tensor()} of shape (1, W, H) for grayscale image or (3, W, H) for
color image to display}

\item{browser}{argument passed to \link{browseURL}}
}
\description{
Display image tensor into browser
}
\seealso{
Other image display:
\code{\link[=draw_bounding_boxes]{draw_bounding_boxes()}},
\code{\link[=draw_keypoints]{draw_keypoints()}},
\code{\link[=draw_segmentation_masks]{draw_segmentation_masks()}},
\code{\link[=tensor_image_display]{tensor_image_display()}},
\code{\link[=vision_make_grid]{vision_make_grid()}}
}
\concept{image display}
