% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-magick.R
\name{tfm_resize}
\alias{tfm_resize}
\title{Resize the input Magick Image to the given size.}
\usage{
tfm_resize(img, size, interpolation = 2)
}
\arguments{
\item{img}{(Magick Image): Image to be resized.}

\item{size}{(sequence or int): Desired output size. If size is a sequence like
(h, w), the output size will be matched to this. If size is an int,
the smaller edge of the image will be matched to this number maintaining
the aspect ratio. i.e, if height > width, then image will be rescaled to
\eqn{\left(\text{size} \times \frac{\text{height}}{\text{width}}, \text{size}\right)}.
For compatibility reasons with \code{tft_resize}, if a tuple or list
of length 1 is provided, it is interpreted as a single int.}

\item{interpolation}{(int, optional): Desired interpolation. An integer \code{0 = nearest},
\code{2 = bilinear} and \code{3 = bicubic} or a name from \code{\link[magick:options]{magick::filter_types()}}.}
}
\value{
Magick Image: Resized image.
}
\description{
Resize the input Magick Image to the given size.
}
