% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_adjust_contrast}
\alias{tft_adjust_contrast}
\title{Adjust contrast of an RGB image.}
\usage{
tft_adjust_contrast(img, contrast_factor)
}
\arguments{
\item{img}{(Tensor): Image to be adjusted.}

\item{contrast_factor}{(float): How much to adjust the contrast. Can be any
non negative number. 0 gives a solid gray image, 1 gives the
original image while 2 increases the contrast by a factor of 2.}
}
\value{
Contrast adjusted image.
}
\description{
Adjust contrast of an RGB image.
}
