% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_adjust_gamma}
\alias{tft_adjust_gamma}
\title{Adjust gamma of an RGB image.}
\usage{
tft_adjust_gamma(img, gamma, gain = 1)
}
\arguments{
\item{img}{(Tensor): Tensor of RBG values to be adjusted.}

\item{gamma}{(float): Non negative real number, same as \eqn{\gamma} in the equation.
gamma larger than 1 make the shadows darker,
while gamma smaller than 1 make dark regions lighter.}

\item{gain}{(float): The constant multiplier.}
}
\description{
Also known as Power Law Transform. Intensities in RGB mode are adjusted
based on the following equation:
}
\details{
\deqn{
  I_{\mbox{out}} = 255 \times \mbox{gain} \times \left(\frac{I_{\mbox{in}}}{255}\right)^{\gamma}
}

See \href{https://en.wikipedia.org/wiki/Gamma_correction}{Gamma Correction}
for more details.
}
