% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_adjust_saturation}
\alias{tft_adjust_saturation}
\title{Adjust color saturation of an RGB image.}
\usage{
tft_adjust_saturation(img, saturation_factor)
}
\arguments{
\item{img}{(Tensor): Image to be adjusted.}

\item{saturation_factor}{(float):  How much to adjust the saturation. Can be any
non negative number. 0 gives a black and white image, 1 gives the
original image while 2 enhances the saturation by a factor of 2.}
}
\value{
Tensor: Saturation adjusted image.
}
\description{
Adjust color saturation of an RGB image.
}
