% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_affine}
\alias{tft_affine}
\title{Apply affine transformation on the Tensor image keeping image center invariant.}
\usage{
tft_affine(img, matrix, resample = 0, fillcolor = NULL)
}
\arguments{
\item{img}{(Tensor): image to be rotated.}

\item{matrix}{(list of floats): list of 6 float values representing inverse matrix for affine transformation.}

\item{resample}{(int, optional): An optional resampling filter. Default is nearest (=2). Other supported values:
bilinear(=2).}

\item{fillcolor}{(int, optional): this option is not supported for Tensor input. Fill value for the area outside the
transform in the output image is always 0.}
}
\value{
Tensor: Transformed image.
}
\description{
Apply affine transformation on the Tensor image keeping image center invariant.
}
