% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_center_crop}
\alias{tft_center_crop}
\title{Crop the Image Tensor and resize it to desired size.}
\usage{
tft_center_crop(img, output_size)
}
\arguments{
\item{img}{(Tensor): Image to be cropped.}

\item{output_size}{(sequence or int): (height, width) of the crop box. If int,
it is used for both directions}
}
\value{
Tensor: Cropped image.
}
\description{
Crop the Image Tensor and resize it to desired size.
}
