% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_crop}
\alias{tft_crop}
\title{Crop the given Image Tensor.}
\usage{
tft_crop(img, top, left, height, width)
}
\arguments{
\item{img}{(Tensor): Image to be cropped in the form \link{..., H, W}. (0,0) denotes the top left corner of the image.}

\item{top}{(int): Vertical component of the top left corner of the crop box.}

\item{left}{(int): Horizontal component of the top left corner of the crop box.}

\item{height}{(int): Height of the crop box.}

\item{width}{(int): Width of the crop box.}
}
\value{
Cropped image.
}
\description{
Crop the given Image Tensor.
}
