% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_hflip}
\alias{tft_hflip}
\title{Horizontally flip the given the Image Tensor.}
\usage{
tft_hflip(img)
}
\arguments{
\item{img}{(Tensor): Image Tensor to be flipped in the form \link{C, H, W}.}
}
\value{
Horizontally flipped image Tensor.
}
\description{
Horizontally flip the given the Image Tensor.
}
