% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_pad}
\alias{tft_pad}
\title{Pad the given Tensor Image on all sides with specified
padding mode and fill value.}
\usage{
tft_pad(img, padding, fill = 0, padding_mode = "constant")
}
\arguments{
\item{img}{(Tensor): Image to be padded.}

\item{padding}{(int or tuple or list): Padding on each border. If a single int is provided this
is used to pad all borders. If a tuple or list of length 2 is provided this is the padding
on left/right and top/bottom respectively. If a tuple or list of length 4 is provided
this is the padding for the left, top, right and bottom borders
respectively. In torchscript mode padding as single int is not supported, use a tuple or
list of length 1: \verb{[padding, ]}.}

\item{fill}{(int): Pixel fill value for constant fill. Default is 0.
This value is only used when the padding_mode is constant}

\item{padding_mode}{(str): Type of padding. Should be: constant, edge or reflect. Default is constant.
Mode symmetric is not yet supported for Tensor inputs.
\itemize{
\item constant: pads with a constant value, this value is specified with fill
\item edge: pads with the last value on the edge of the image
\item reflect: pads with reflection of image (without repeating the last value on the edge)
padding \verb{[1, 2, 3, 4]} with 2 elements on both sides in reflect mode
will result in \verb{[3, 2, 1, 2, 3, 4, 3, 2]}
\item symmetric: pads with reflection of image (repeating the last value on the edge)
padding \verb{[1, 2, 3, 4]} with 2 elements on both sides in symmetric mode
will result in \verb{[2, 1, 1, 2, 3, 4, 4, 3]}
}}
}
\value{
Tensor: Padded image.
}
\description{
Pad the given Tensor Image on all sides with specified
padding mode and fill value.
}
