% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_resize}
\alias{tft_resize}
\title{Resize the input Tensor to the given size.}
\usage{
tft_resize(img, size, interpolation = 2)
}
\arguments{
\item{img}{(Tensor): Image to be resized.}

\item{size}{(int or tuple or list): Desired output size. If size is a sequence like
(h, w), the output size will be matched to this. If size is an int,
the smaller edge of the image will be matched to this number maintaining
the aspect ratio. i.e, if height > width, then image will be rescaled to
\eqn{\left(\text{size} \times \frac{\text{height}}{\text{width}}, \text{size}\right)}.}

\item{interpolation}{(int, optional): Desired interpolation. Default is bilinear (=2). Other supported values:
nearest(=0) and bicubic(=3).}
}
\value{
Tensor: Resized image.
}
\description{
Resize the input Tensor to the given size.
}
