% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_rgb_to_grayscale}
\alias{tft_rgb_to_grayscale}
\title{Convert the given RGB Image Tensor to Grayscale.
For RGB to Grayscale conversion, ITU-R 601-2 luma transform is performed which
is L = R * 0.2989 + G * 0.5870 + B * 0.1140}
\usage{
tft_rgb_to_grayscale(img)
}
\arguments{
\item{img}{(Tensor): Image to be converted to Grayscale in the
form \verb{[C, H, W]}.}
}
\value{
Grayscale image.
}
\description{
Convert the given RGB Image Tensor to Grayscale.
For RGB to Grayscale conversion, ITU-R 601-2 luma transform is performed which
is L = R * 0.2989 + G * 0.5870 + B * 0.1140
}
