% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_ten_crop}
\alias{tft_ten_crop}
\title{Crop the given Image Tensor into four corners and the central crop plus the
flipped version of these (horizontal flipping is used by default).}
\usage{
tft_ten_crop(img, size, vertical_flip = FALSE)
}
\arguments{
\item{img}{(Tensor): Image to be cropped.}

\item{size}{(sequence or int): Desired output size of the crop. If size is an
int instead of sequence like (h, w), a square crop (size, size) is
made.}

\item{vertical_flip}{(bool): Use vertical flipping instead of horizontal}
}
\value{
List: List (tl, tr, bl, br, center, tl_flip, tr_flip, bl_flip, br_flip, center_flip)
Corresponding top left, top right, bottom left, bottom right and center crop
and same for the flipped image's tensor.
}
\description{
Crop the given Image Tensor into four corners and the central crop plus the
flipped version of these (horizontal flipping is used by default).
}
\note{
This transform returns a List of images and there may be a
mismatch in the number of inputs and targets your \code{Dataset} returns.
}
