% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-functional-tensor.R
\name{tft_vflip}
\alias{tft_vflip}
\title{Vertically flip the given the Image Tensor.}
\usage{
tft_vflip(img)
}
\arguments{
\item{img}{(Tensor): Image Tensor to be flipped in the form \verb{[C, H, W]}.}
}
\value{
Vertically flipped image Tensor.
}
\description{
Vertically flip the given the Image Tensor.
}
