% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-imagenet.R
\name{tiny_imagenet_dataset}
\alias{tiny_imagenet_dataset}
\title{Tiny ImageNet dataset}
\usage{
tiny_imagenet_dataset(root, split = "train", download = FALSE, ...)
}
\arguments{
\item{root}{directory path to download the dataset.}

\item{split}{dataset split, \code{train}, \code{validation} or \code{test}.}

\item{download}{whether to download or not the dataset.}

\item{...}{other arguments passed to \code{\link[=image_folder_dataset]{image_folder_dataset()}}.}
}
\description{
Prepares the Tiny ImageNet dataset and optionally downloads it.
}
\seealso{
Other classification_dataset:
\code{\link{caltech_dataset}},
\code{\link[=cifar10_dataset]{cifar10_dataset()}},
\code{\link[=eurosat_dataset]{eurosat_dataset()}},
\code{\link[=fer_dataset]{fer_dataset()}},
\code{\link[=fgvc_aircraft_dataset]{fgvc_aircraft_dataset()}},
\code{\link[=flowers102_dataset]{flowers102_dataset()}},
\code{\link[=image_folder_dataset]{image_folder_dataset()}},
\code{\link{lfw_dataset}},
\code{\link[=mnist_dataset]{mnist_dataset()}},
\code{\link[=oxfordiiitpet_dataset]{oxfordiiitpet_dataset()}},
\code{\link[=places365_dataset]{places365_dataset()}},
\code{\link[=vggface2_dataset]{vggface2_dataset()}},
\code{\link[=whoi_plankton_dataset]{whoi_plankton_dataset()}},
\code{\link[=whoi_small_coralnet_dataset]{whoi_small_coralnet_dataset()}}
}
\concept{classification_dataset}
