% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_utils.R
\name{torch_browse}
\alias{torch_browse}
\title{Display image tensor}
\usage{
torch_browse(image, browser = getOption("browser"))
}
\arguments{
\item{image}{\code{torch_tensor()} of shape (1, W, H) for grayscale image or (3, W, H) for color image,
of type \code{torch_uint8()} to display}

\item{browser}{argument passed to \link{browseURL}}
}
\description{
Display image tensor into browser
}
