% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_utils.R
\name{torch_display}
\alias{torch_display}
\title{Display image tensor}
\usage{
torch_display(image, animate = TRUE)
}
\arguments{
\item{image}{\code{torch_tensor()} of shape (1, W, H) for grayscale image or (3, W, H) for color image,
of type \code{torch_uint8()} to display}

\item{animate}{support animations in the X11 display}
}
\description{
Display image tensor onto the X11 device
}
