% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spam_model.R
\name{train_model}
\alias{train_model}
\title{Train a neural network model}
\usage{
train_model(model, train_dl, test_dl, lr = 0.001, epochs = 10, ...)
}
\arguments{
\item{model}{The torch model}

\item{train_dl}{Training dataloader}

\item{test_dl}{Test dataloader (not used in training)}

\item{lr}{Learning rate}

\item{epochs}{Number of epochs}
}
\value{
Trained model
}
\description{
Train a neural network model
}
