% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_adjust_brightness}
\alias{transform_adjust_brightness}
\title{Adjust brightness of an Image.}
\usage{
transform_adjust_brightness(img, brightness_factor)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be adjusted.}

\item{brightness_factor}{(float):  How much to adjust the brightness. Can be
any non negative number. 0 gives a black image, 1 gives the
original image while 2 increases the brightness by a factor of 2.}
}
\value{
Magick Image or Tensor: Brightness adjusted image.
}
\description{
Adjust brightness of an Image.
}
