% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_adjust_gamma}
\alias{transform_adjust_gamma}
\title{Perform gamma correction on an image.}
\usage{
transform_adjust_gamma(img, gamma, gain = 1)
}
\arguments{
\item{img}{(Magick Image or Tensor): PIL Image to be adjusted.}

\item{gamma}{(float): Non negative real number, same as \eqn{\gamma} in the equation.
gamma larger than 1 make the shadows darker,
while gamma smaller than 1 make dark regions lighter.}

\item{gain}{(float): The constant multiplier.}
}
\value{
Magick Image or Tensor: Gamma correction adjusted image.
}
\description{
Also known as Power Law Transform. Intensities in RGB mode are adjusted
based on the following equation:
}
\details{
\eqn{
 I_{\text{out}} = 255 \times \text{gain} \times \left(\frac{I_{\text{in}}}{255}\right)^{\gamma}
}

See \href{https://en.wikipedia.org/wiki/Gamma_correction}{Gamma Correction} for
more details.
}
