% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-generics.R
\name{transform_convert_image_dtype}
\alias{transform_convert_image_dtype}
\title{Convert a tensor image to the given \code{dtype} and scale the values accordingly}
\usage{
transform_convert_image_dtype(img, dtype = torch::torch_float())
}
\arguments{
\item{img}{A \code{magick-image}, \code{array} or \code{torch_tensor}.}

\item{dtype}{(torch.dtype): Desired data type of the output.}
}
\description{
Convert a tensor image to the given \code{dtype} and scale the values accordingly
}
\note{
When converting from a smaller to a larger integer \code{dtype} the maximum
values are \strong{not} mapped exactly. If converted back and forth, this
mismatch has no effect.
}
\seealso{
Other unitary_transforms: 
\code{\link{transform_adjust_brightness}()},
\code{\link{transform_adjust_contrast}()},
\code{\link{transform_adjust_gamma}()},
\code{\link{transform_adjust_hue}()},
\code{\link{transform_adjust_saturation}()},
\code{\link{transform_affine}()},
\code{\link{transform_center_crop}()},
\code{\link{transform_crop}()},
\code{\link{transform_hflip}()},
\code{\link{transform_linear_transformation}()},
\code{\link{transform_normalize}()},
\code{\link{transform_pad}()},
\code{\link{transform_perspective}()},
\code{\link{transform_resize}()},
\code{\link{transform_rgb_to_grayscale}()},
\code{\link{transform_rotate}()},
\code{\link{transform_to_tensor}()},
\code{\link{transform_vflip}()}
}
\concept{unitary_transforms}
