% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_crop}
\alias{transform_crop}
\title{Crop the given image at specified location and output size.}
\usage{
transform_crop(img, top, left, height, width)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be cropped. (0,0) denotes the top left corner of the image.}

\item{top}{(int): Vertical component of the top left corner of the crop box.}

\item{left}{(int): Horizontal component of the top left corner of the crop box.}

\item{height}{(int): Height of the crop box.}

\item{width}{(int): Width of the crop box.}
}
\value{
Magick Image or Tensor: Cropped image.
}
\description{
The image can be a Magick Image or a Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading
dimensions
}
