% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_five_crop}
\alias{transform_five_crop}
\title{Crop the given image into four corners and the central crop.}
\usage{
transform_five_crop(img, size)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be cropped.}

\item{size}{(sequence or int): Desired output size of the crop. If size is an
int instead of sequence like (h, w), a square crop (size, size) is
made. If provided a tuple or list of length 1, it will be interpreted as
\verb{(size[1], size[2])}.}
}
\value{
tuple: tuple (tl, tr, bl, br, center)
Corresponding top left, top right, bottom left, bottom right and center crop.
}
\description{
The image can be a Magick Image or a Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading
dimensions.
}
\note{
This transform returns a tuple of images and there may be a
mismatch in the number of inputs and targets your \code{Dataset} returns.
}
