% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-generics.R
\name{transform_grayscale}
\alias{transform_grayscale}
\title{Convert image to grayscale}
\usage{
transform_grayscale(img, num_output_channels)
}
\arguments{
\item{img}{A \code{magick-image}, \code{array} or \code{torch_tensor}.}

\item{num_output_channels}{(int): (1 or 3) number of channels desired for
output image}
}
\description{
Convert image to grayscale
}
\seealso{
Other unitary_transforms:
\code{\link[=transform_adjust_brightness]{transform_adjust_brightness()}},
\code{\link[=transform_adjust_contrast]{transform_adjust_contrast()}},
\code{\link[=transform_adjust_gamma]{transform_adjust_gamma()}},
\code{\link[=transform_adjust_hue]{transform_adjust_hue()}},
\code{\link[=transform_adjust_saturation]{transform_adjust_saturation()}},
\code{\link[=transform_affine]{transform_affine()}},
\code{\link[=transform_center_crop]{transform_center_crop()}},
\code{\link[=transform_convert_image_dtype]{transform_convert_image_dtype()}},
\code{\link[=transform_crop]{transform_crop()}},
\code{\link[=transform_hflip]{transform_hflip()}},
\code{\link[=transform_linear_transformation]{transform_linear_transformation()}},
\code{\link[=transform_normalize]{transform_normalize()}},
\code{\link[=transform_pad]{transform_pad()}},
\code{\link[=transform_perspective]{transform_perspective()}},
\code{\link[=transform_resize]{transform_resize()}},
\code{\link[=transform_rgb_to_grayscale]{transform_rgb_to_grayscale()}},
\code{\link[=transform_rotate]{transform_rotate()}},
\code{\link[=transform_to_tensor]{transform_to_tensor()}},
\code{\link[=transform_vflip]{transform_vflip()}}
}
\concept{unitary_transforms}
