% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_hflip}
\alias{transform_hflip}
\title{Horizontally flip the given PIL Image or Tensor.}
\usage{
transform_hflip(img)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be flipped. If img
is a Tensor, it is expected to be in \verb{[..., H, W]} format,
where ... means it can have an arbitrary number of trailing
dimensions.}
}
\value{
Magick Image or Tensor:  Horizontally flipped image.
}
\description{
Horizontally flip the given PIL Image or Tensor.
}
