% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_magick_to_tensor}
\alias{transform_magick_to_tensor}
\title{Convert a \verb{PIL Image} to a tensor of the same type.}
\usage{
transform_magick_to_tensor(pic)
}
\arguments{
\item{pic}{(PIL Image): Image to be converted to tensor.}
}
\value{
Tensor: Converted image.
}
\description{
Convert a \verb{PIL Image} to a tensor of the same type.
}
