% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_normalize}
\alias{transform_normalize}
\title{Normalize a tensor image with mean and standard deviation.}
\usage{
transform_normalize(tensor, mean, std, inplace = FALSE)
}
\arguments{
\item{tensor}{(Tensor): Tensor image of size (C, H, W) to be normalized.}

\item{mean}{(sequence): Sequence of means for each channel.}

\item{std}{(sequence): Sequence of standard deviations for each channel.}

\item{inplace}{(bool,optional): Bool to make this operation inplace.}
}
\value{
Tensor: Normalized Tensor image.
}
\description{
Normalize a tensor image with mean and standard deviation.
}
\note{
This transform acts out of place by default, i.e., it does not mutates the
input tensor.
}
