% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_perspective}
\alias{transform_perspective}
\title{Perform perspective transform of the given image.}
\usage{
transform_perspective(
  img,
  startpoints,
  endpoints,
  interpolation = 2,
  fill = NULL
)
}
\arguments{
\item{img}{(PIL Image or Tensor): Image to be transformed.}

\item{startpoints}{(list of list of ints): List containing four lists of two integers corresponding to four corners
\verb{[top-left, top-right, bottom-right, bottom-left]} of the original image.}

\item{endpoints}{(list of list of ints): List containing four lists of two integers corresponding to four corners
\verb{[top-left, top-right, bottom-right, bottom-left]} of the transformed image.}

\item{interpolation}{(int, optional): Desired interpolation enum defined by \code{filters}.
Default is \code{2 = BILINEAR}. If input is Tensor, only \code{0 = NEAREST}, \code{2 = BILINEAR}
and \code{3 = BICUBIC} are supported.}

\item{fill}{(int or str or tuple): Pixel fill value for constant fill. Default is 0. If a tuple of
length 3, it is used to fill R, G, B channels respectively.
This value is only used when the padding_mode is constant. Only int value is
supported for Tensors.}
}
\value{
Magick Image or Tensor: transformed Image.
}
\description{
The image can be a PIL Image or a Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading
dimensions.
}
