% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-generics.R
\name{transform_random_apply}
\alias{transform_random_apply}
\title{Apply a list of transformations randomly with a given probability}
\usage{
transform_random_apply(img, transforms, p = 0.5)
}
\arguments{
\item{img}{A \code{magick-image}, \code{array} or \code{torch_tensor}.}

\item{transforms}{(list or tuple): list of transformations.}

\item{p}{(float): probability.}
}
\description{
Apply a list of transformations randomly with a given probability
}
\seealso{
Other combining_transforms: 
\code{\link{transform_five_crop}()},
\code{\link{transform_random_choice}()},
\code{\link{transform_random_order}()},
\code{\link{transform_resized_crop}()},
\code{\link{transform_ten_crop}()}
}
\concept{combining_transforms}
