% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-generics.R
\name{transform_random_choice}
\alias{transform_random_choice}
\title{Apply single transformation randomly picked from a list}
\usage{
transform_random_choice(img, transforms)
}
\arguments{
\item{img}{A \code{magick-image}, \code{array} or \code{torch_tensor}.}

\item{transforms}{(list or tuple): list of transformations.}
}
\description{
Apply single transformation randomly picked from a list
}
\seealso{
Other combining_transforms: 
\code{\link{transform_five_crop}()},
\code{\link{transform_random_apply}()},
\code{\link{transform_random_order}()},
\code{\link{transform_resized_crop}()},
\code{\link{transform_ten_crop}()}
}
\concept{combining_transforms}
