% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-generics.R
\name{transform_rgb_to_grayscale}
\alias{transform_rgb_to_grayscale}
\title{Convert RGB Image Tensor to Grayscale}
\usage{
transform_rgb_to_grayscale(img)
}
\arguments{
\item{img}{A \code{magick-image}, \code{array} or \code{torch_tensor}.}
}
\description{
For RGB to Grayscale conversion, ITU-R 601-2 luma transform is performed
which is L = R * 0.2989 + G * 0.5870 + B * 0.1140
}
\seealso{
Other unitary_transforms: 
\code{\link{transform_adjust_brightness}()},
\code{\link{transform_adjust_contrast}()},
\code{\link{transform_adjust_gamma}()},
\code{\link{transform_adjust_hue}()},
\code{\link{transform_adjust_saturation}()},
\code{\link{transform_affine}()},
\code{\link{transform_center_crop}()},
\code{\link{transform_convert_image_dtype}()},
\code{\link{transform_crop}()},
\code{\link{transform_grayscale}()},
\code{\link{transform_hflip}()},
\code{\link{transform_linear_transformation}()},
\code{\link{transform_normalize}()},
\code{\link{transform_pad}()},
\code{\link{transform_perspective}()},
\code{\link{transform_resize}()},
\code{\link{transform_rotate}()},
\code{\link{transform_to_tensor}()},
\code{\link{transform_vflip}()}
}
\concept{unitary_transforms}
