% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_ten_crop}
\alias{transform_ten_crop}
\title{Generate ten cropped images from the given image.}
\usage{
transform_ten_crop(img, size, vertical_flip = FALSE)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be cropped.}

\item{size}{(sequence or int): Desired output size of the crop. If size is an
int instead of sequence like (h, w), a square crop (size, size) is
made. If provided a tuple or list of length 1, it will be interpreted as \verb{(size[1], size[2])}.}

\item{vertical_flip}{(bool): Use vertical flipping instead of horizontal}
}
\value{
tuple: tuple (tl, tr, bl, br, center, tl_flip, tr_flip, bl_flip, br_flip, center_flip)
Corresponding top left, top right, bottom left, bottom right and
center crop and same for the flipped image.
}
\description{
Crop the given image into four corners and the central crop plus the
flipped version of these (horizontal flipping is used by default).
The image can be a PIL Image or a Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading
dimensions
}
\note{
This transform returns a tuple of images and there may be a
mismatch in the number of inputs and targets your \code{Dataset} returns.
}
