% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_to_tensor}
\alias{transform_to_tensor}
\title{Convert a \verb{Magick Image} or \code{array} to tensor.}
\usage{
transform_to_tensor(pic)
}
\arguments{
\item{pic}{(\verb{Magick Image} or \code{array}): Image to be converted to tensor.}
}
\value{
Tensor: Converted image.
}
\description{
Convert a \verb{Magick Image} or \code{array} to tensor.
}
