% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_vflip}
\alias{transform_vflip}
\title{Vertically flip the given PIL Image or torch Tensor.}
\usage{
transform_vflip(img)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be flipped. If img
is a Tensor, it is expected to be in \verb{[..., H, W]} format,
where ... means it can have an arbitrary number of trailing
dimensions.}
}
\value{
Magick Image:  Vertically flipped image.
}
\description{
Vertically flip the given PIL Image or torch Tensor.
}
