% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-vggface2.R
\name{vggface2_dataset}
\alias{vggface2_dataset}
\title{VGGFace2 Dataset}
\usage{
vggface2_dataset(
  root = tempdir(),
  split = "val",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Character. Root directory where the dataset will be stored under \code{root/vggface2}.}

\item{split}{One of \code{"train"}, \code{"val"}, or \code{"test"}. Default is \code{"train"}.}

\item{transform}{Optional function to transform input images after loading. Default is \code{NULL}.}

\item{target_transform}{Optional function to transform labels. Default is \code{NULL}.}

\item{download}{Logical. Whether to download the dataset if not found locally. Default is \code{FALSE}.}
}
\value{
A torch dataset object \code{vggface2_dataset}:
\itemize{
\item \code{x}: RGB image array.
\item \code{y}: Integer label (1…N) for the identity.
}

\code{ds$classes} is a named list mapping integer labels to a list with:
\itemize{
\item \code{name}: Character name of the person.
\item \code{gender}: "Male" or "Female".
}
}
\description{
The VGGFace2 dataset is a large-scale face recognition dataset containing images
of celebrities from a wide range of ethnicities, professions, and ages.
Each identity has multiple images with large variations in pose, age, illumination,
ethnicity, and profession.
}
\examples{
\dontrun{
#Load the training set
ds <- vggface2_dataset(download = TRUE)
item <- ds[1]
item$x      # image array RGB
item$y      # integer label
ds$classes[item$y]  # list(name=..., gender=...)

#Load the test set
ds <- vggface2_dataset(download = TRUE, train = FALSE)
item <- ds[1]
item$x      # image array RGB
item$y      # integer label
ds$classes[item$y]  # list(name=..., gender=...)
}

}
\seealso{
Other segmentation_dataset: 
\code{\link{oxfordiiitpet_segmentation_dataset}()},
\code{\link{pascal_voc_datasets}}
}
\concept{segmentation_dataset}
