% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-deeplabv3.R
\docType{data}
\name{voc_classes}
\alias{voc_classes}
\alias{voc_label}
\title{PASCAL VOC Class Labels}
\format{
An object of class \code{character} of length 21.
}
\usage{
voc_classes

voc_label(id)
}
\arguments{
\item{id}{Integer vector of 1-based class identifiers.}
}
\value{
A character vector with the PASCAL VOC class names

A character vector with the labels associated with \code{id}.
}
\description{
Utilities for resolving PASCAL VOC class identifiers to their corresponding
human readable labels. The labels are retrieved from the dataset.
}
\seealso{
Other class_resolution: 
\code{\link{coco_classes}()},
\code{\link{imagenet_classes}()},
\code{\link{voc_segmentation_classes}}

Other class_resolution: 
\code{\link{coco_classes}()},
\code{\link{imagenet_classes}()},
\code{\link{voc_segmentation_classes}}
}
\concept{class_resolution}
\keyword{datasets}
