% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-fcn.R
\docType{data}
\name{voc_segmentation_classes}
\alias{voc_segmentation_classes}
\title{PASCAL-VOC Class Labels}
\format{
An object of class \code{character} of length 21.
}
\usage{
voc_segmentation_classes
}
\value{
A character vector with 21 entries representing the PASCAL-VOC
class labels.
}
\description{
PASCAL-VOC Class Labels
}
\seealso{
Other class_resolution: 
\code{\link{coco_classes}()},
\code{\link{imagenet_classes}()},
\code{\link{voc_classes}}
}
\concept{class_resolution}
\keyword{datasets}
