% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-plankton.R
\name{whoi_plankton_dataset}
\alias{whoi_plankton_dataset}
\alias{whoi_small_plankton_dataset}
\title{WHOI Plankton Datasets}
\usage{
whoi_small_plankton_dataset(
  split = "val",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)

whoi_plankton_dataset(
  split = "val",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{split}{One of \code{"train"}, \code{"val"}, or \code{"test"}. Default is \code{"train"}.}

\item{transform}{Optional. A function that takes an image and returns a transformed version (e.g., normalization, cropping).}

\item{target_transform}{Optional. A function that transforms the label.}

\item{download}{Logical. If TRUE, downloads the dataset to \verb{root/}. If the dataset is already present, download is skipped.}
}
\value{
A torch dataset.
Each element is a named list:
\itemize{
\item \code{x}: a H x W x 1 integer array representing an grayscale image.
\item \code{y}: the class id of the image.
}
}
\description{
WHOI-Plankton Dataset
}
\details{
The WHOI-Plankton and WHOI-Plankton small are \strong{image classification} datasets
composed of 957k and 58k grayscale images respectively, within classes
The images are in format with varying spatial resolutions.
}
\examples{
\dontrun{
# Load the small plankton dataset
plankton <- whoi_small_plankton_dataset(download = TRUE)

# Access the first item
first_item <- plankton[1]
first_item$x  # image array with shape {H, W, 1}
first_item$y  # id of the plankton class.

# Load the full plankton dataset
plankton <- whoi_plankton_dataset(download = TRUE)

# Access the first item
first_item <- plankton[1]
first_item$x  # image array with shape {H, W, 1}
first_item$y  # id of the plankton class.
}

}
\seealso{
Other classification_dataset: 
\code{\link{caltech_dataset}},
\code{\link{cifar10_dataset}()},
\code{\link{eurosat_dataset}()},
\code{\link{fer_dataset}()},
\code{\link{fgvc_aircraft_dataset}()},
\code{\link{flowers102_dataset}()},
\code{\link{lfw_dataset}},
\code{\link{mnist_dataset}()},
\code{\link{oxfordiiitpet_dataset}()},
\code{\link{places365_dataset}()},
\code{\link{tiny_imagenet_dataset}()}
}
\concept{classification_dataset}
