% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-plankton.R
\name{whoi_small_coralnet_dataset}
\alias{whoi_small_coralnet_dataset}
\title{Coralnet Dataset}
\usage{
whoi_small_coralnet_dataset(
  split = "val",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{split}{One of \code{"train"}, \code{"val"}, or \code{"test"}. Default is \code{"val"}.}

\item{transform}{Optional. A function that takes an image and returns a transformed version (e.g., normalization, cropping).}

\item{target_transform}{Optional. A function that transforms the label.}

\item{download}{Logical. If TRUE, downloads the dataset to \verb{root/}. If the dataset is already present, download is skipped.}
}
\description{
Small Coralnet dataset is an image \strong{classification dataset}
of very large submarine coral reef images annotated into 3 classes
and produced by  \href{https://coralnet.ucsd.edu}{CoralNet},
a resource for benthic images classification.
}
\seealso{
Other classification_dataset: 
\code{\link{caltech_dataset}},
\code{\link{cifar10_dataset}()},
\code{\link{eurosat_dataset}()},
\code{\link{fer_dataset}()},
\code{\link{fgvc_aircraft_dataset}()},
\code{\link{flowers102_dataset}()},
\code{\link{image_folder_dataset}()},
\code{\link{lfw_dataset}},
\code{\link{mnist_dataset}()},
\code{\link{oxfordiiitpet_dataset}()},
\code{\link{places365_dataset}()},
\code{\link{tiny_imagenet_dataset}()},
\code{\link{vggface2_dataset}()},
\code{\link{whoi_plankton_dataset}()}
}
\concept{classification_dataset}
