% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-boxes.R
\name{box_convert}
\alias{box_convert}
\title{Box Convert}
\usage{
box_convert(boxes, in_fmt, out_fmt)
}
\arguments{
\item{boxes}{(Tensor\link{N, 4}): boxes which will be converted.}

\item{in_fmt}{(str): Input format of given boxes. Supported formats are \link{'xyxy', 'xywh', 'cxcywh'}.}

\item{out_fmt}{(str): Output format of given boxes. Supported formats are \link{'xyxy', 'xywh', 'cxcywh'}}
}
\value{
boxes (Tensor\link{N, 4}): Boxes into converted format.
}
\description{
Converts boxes from given in_fmt to out_fmt.
}
\details{
Supported in_fmt and out_fmt are:
'xyxy': boxes are represented via corners, x1, y1 being top left and x2, y2 being bottom right.
'xywh' : boxes are represented via corner, width and height, x1, y2 being top left, w, h being width and height.
'cxcywh' : boxes are represented via centre, width and height, cx, cy being center of box, w, h
being width and height.
}
