% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_cxcywh_to_xyxy}
\alias{box_cxcywh_to_xyxy}
\title{box_cxcywh_to_xyxy}
\usage{
box_cxcywh_to_xyxy(boxes)
}
\arguments{
\item{boxes}{(Tensor\link{N, 4}): boxes in (cx, cy, w, h) format which will be converted.}
}
\value{
boxes (Tensor(N, 4)): boxes in (x1, y1, x2, y2) format.
}
\description{
Converts bounding boxes from  (cx, cy, w, h) format to (x1, y1, x2, y2) format.
(cx, cy) refers to center of bounding box
(w, h) are width and height of bounding box
}
