% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-boxes.R
\name{box_iou}
\alias{box_iou}
\title{Box IoU}
\usage{
box_iou(boxes1, boxes2)
}
\arguments{
\item{boxes1}{(Tensor[N, 4])}

\item{boxes2}{(Tensor[M, 4])}
}
\value{
iou (Tensor[N, M]): the NxM matrix containing the pairwise IoU values for every element in boxes1 and boxes2
}
\description{
Return intersection-over-union  (Jaccard index) of boxes.
Both sets of boxes are expected to be in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format with
\eqn{0 \leq x_{min} < x_{max}} and \eqn{0 \leq y_{min} < y_{max}}.
}
