% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_xyxy_to_cxcywh}
\alias{box_xyxy_to_cxcywh}
\title{box_xyxy_to_cxcywh}
\usage{
box_xyxy_to_cxcywh(boxes)
}
\value{
boxes (Tensor(N, 4)): boxes in (cx, cy, w, h) format.
}
\description{
Converts bounding boxes from  (x1, y1, x2, y2) format to (cx, cy, w, h) format.
(x1, y1) refer to top left of bounding box
(x2, y2) refer to bottom right of bounding box
}
\details{
@param boxes  (Tensor\link{N, 4}): boxes in (x1, y1, x2, y2) format which will be converted.
}
