% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-box_convert.R
\name{box_xyxy_to_xyxyr}
\alias{box_xyxy_to_xyxyr}
\title{box_xyxy_to_xyxyr}
\usage{
box_xyxy_to_xyxyr(boxes)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format
which will be converted.}
}
\value{
boxes (Tensor[N, 5]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max}, r)} format,
with rotation angle \eqn{r = 0}.
}
\description{
Converts bounding boxes from \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format to
\eqn{(x_{min}, y_{min}, x_{max}, y_{max}, r)} format, where \eqn{r} is the rotation
angle in radians (anti-clockwise). For axis-aligned boxes, \eqn{r = 0}.
}
