% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-caltech.R
\name{caltech101_dataset}
\alias{caltech101_dataset}
\title{Caltech-101 Dataset Loader (from Local Storage)}
\usage{
caltech101_dataset(root, split = "train")
}
\arguments{
\item{root}{Character. The root directory where the dataset is stored.}

\item{split}{Character. One of \code{train} or \code{test}.}

\item{transform}{Function. Optional image transformation.}

\item{target_transform}{Function. Optional label transformation.}
}
\value{
An R6 dataset object that inherits from \code{torch::dataset}.
}
\description{
Loads the Caltech-101 dataset stored in the 'data/' folder.
}
