% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-boxes.R
\name{clip_boxes_to_image}
\alias{clip_boxes_to_image}
\title{Clip Boxes to Image}
\usage{
clip_boxes_to_image(boxes, size)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format
with
\itemize{
\item \eqn{0 \leq x_{min} < x_{max}} and
\item \eqn{0 \leq y_{min} < y_{max}}.
}}

\item{size}{(Tuple[height, width]): size of the image}
}
\value{
clipped_boxes (Tensor[N, 4])
}
\description{
Clip boxes so that they lie inside an image of size \code{size}.
}
