% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-coco.R
\name{coco_caption_dataset}
\alias{coco_caption_dataset}
\title{COCO Caption Dataset}
\usage{
coco_caption_dataset(root, train = TRUE, year = c("2014"), download = FALSE)
}
\arguments{
\item{root}{Root directory where the dataset is stored or will be downloaded to.}

\item{train}{Logical. If TRUE, loads the training split; otherwise, loads the validation split.}

\item{year}{Character. Dataset version year. One of \code{"2014"}.}

\item{download}{Logical. If TRUE, downloads the dataset if it's not already present in the \code{root} directory.}
}
\description{
Loads the MS COCO dataset for image captioning.
}
\examples{
\dontrun{
ds <- coco_dataset(
  root = "~/data",
  train = FALSE,
  download = TRUE
)
example <- ds[1]

# Access image and caption
x <- example$x
y <- example$y

# Prepare image for plotting
image_array <- as.numeric(x)
dim(image_array) <- dim(x)

plot(as.raster(image_array))
title(main = y, col.main = "black")
}
}
