% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-coco.R
\name{coco_classes}
\alias{coco_classes}
\alias{coco_label}
\title{COCO Class Labels}
\usage{
coco_classes()

coco_label(id)
}
\arguments{
\item{id}{Integer vector of 1-based class identifiers.}
}
\value{
A character vector with the COCO class names

A character vector with the labels associated with \code{id}.
}
\description{
Utilities for resolving COCO 80 class identifiers to their corresponding
human readable labels. The labels are retrieved from ultralytics source
}
\seealso{
Other class_resolution: 
\code{\link{imagenet_21k_classes}()},
\code{\link{imagenet_classes}()},
\code{\link{voc_classes}}

Other class_resolution: 
\code{\link{imagenet_21k_classes}()},
\code{\link{imagenet_classes}()},
\code{\link{voc_classes}}
}
\concept{class_resolution}
