% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_utils.R
\name{draw_bounding_boxes}
\alias{draw_bounding_boxes}
\title{Draw bounding boxes on image}
\usage{
draw_bounding_boxes(x, ...)
}
\arguments{
\item{x}{Tensor (C x H x W) uint8 or float (\link{0,1})}

\item{boxes}{Tensor (N x 4) in c(xmin, ymin, xmax, ymax)}

\item{labels}{character vector of box labels}

\item{colors}{character vector or single color}

\item{fill}{whether to fill the box}

\item{width}{width of the box border}

\item{font}{font family vector}

\item{font_size}{font size in points}
}
\value{
torch_tensor (C x H x W) with boxes drawn
}
\description{
Draw bounding boxes on image
}
\seealso{
Other image display: 
\code{\link{draw_keypoints}()},
\code{\link{vision_make_grid}()}
}
\concept{image display}
