% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_utils.R
\name{draw_keypoints}
\alias{draw_keypoints}
\title{Draws Keypoints}
\usage{
draw_keypoints(
  image,
  keypoints,
  connectivity = NULL,
  colors = NULL,
  radius = 2,
  width = 3
)
}
\arguments{
\item{image}{: Tensor of shape (3, H, W) and dtype uint8}

\item{keypoints}{: Tensor of shape (num_instances, K, 2) the K keypoints location for each of the N instances,}

\item{connectivity}{: Vector of pair of keypoints to be connected}

\item{colors}{: (optional): a viridisMap() dataframe or equivalent with either one row or}

\item{radius}{: radius of the plotted keypoint.}

\item{width}{: width of line connecting keypoints.}
}
\value{
Image Tensor of dtype uint8 with keypoints drawn.
}
\description{
Draws Keypoints on given RGB tensor image.
}
