% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_utils.R
\name{draw_segmentation_masks}
\alias{draw_segmentation_masks}
\title{Draw segmentation masks}
\usage{
draw_segmentation_masks(image, masks, alpha = 0.8, colors = NULL)
}
\arguments{
\item{image}{: torch_tensor of shape (3, H, W) and dtype uint8.}

\item{masks}{: torch_tensor of shape (num_masks, H, W) or (H, W) and dtype bool.}

\item{alpha}{: number between 0 and 1 denoting the transparency of the masks.}

\item{colors}{: (optional): a viridisMap() dataframe or equivalent with either one row or}
}
\value{
torch_tensor of shape (3, H, W) and dtype uint8 of the image with segmentation masks drawn on top.
}
\description{
Draw segmentation masks with their respective colors on top of a given RGB tensor image
}
