% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-mnist.R
\name{emnist_dataset}
\alias{emnist_dataset}
\title{EMNIST dataset}
\usage{
emnist_dataset(
  root,
  split = "balanced",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Character. Root directory for dataset storage (default folder: \verb{root/emnist/processed/}).}

\item{split}{Character. Dataset split to use. One of \code{"byclass"}, \code{"bymerge"}, \code{"balanced"}, \code{"letters"}, \code{"digits"}, or \code{"mnist"}. Default is \code{"balanced"}.}

\item{transform}{Optional function to transform input images.}

\item{target_transform}{Optional function to transform labels.}

\item{download}{Logical. Whether to download the dataset if it is not found locally. Default is \code{FALSE}.}
}
\value{
An EMNIST dataset object.
}
\description{
EMNIST Dataset
}
\details{
Loads the EMNIST dataset, a set of handwritten digits and letters with multiple splits:
\itemize{
\item "byclass": 62 classes (digits + uppercase + lowercase)
\item "bymerge": 47 classes (merged uppercase and lowercase letters)
\item "balanced": 47 classes balanced between digits and letters
\item "letters": 26 letter classes only
\item "digits": 10 digit classes only
\item "mnist": classic 10 digit classes like the original MNIST dataset
}
}
\examples{
\dontrun{
emnist <- emnist_dataset(split = "balanced", download = TRUE)
first_item <- emnist[1]
# image in item 1
first_item$x
# label of item 1
first_item$y
}

}
\seealso{
\code{\link[=mnist_dataset]{mnist_dataset()}}, \code{\link[=kmnist_dataset]{kmnist_dataset()}}, \code{\link[=fashion_mnist_dataset]{fashion_mnist_dataset()}}
}
