% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-fashionmnist.R
\name{fashion_mnist_dataset}
\alias{fashion_mnist_dataset}
\title{FashionMNIST dataset}
\usage{
fashion_mnist_dataset(
  root,
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{(string): Root directory of dataset.}

\item{train}{(bool, optional): If TRUE, creates dataset from \code{training.rds}, otherwise from \code{test.rds}.}

\item{transform}{(callable, optional): Function to transform input data.}

\item{target_transform}{(callable, optional): Function to transform target labels.}

\item{download}{(bool, optional): If TRUE, downloads the dataset from the internet.}
}
\description{
Prepares the FashionMNIST dataset, similar to \code{mnist1_dataset}.
}
