% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-fer.R
\name{fer_dataset}
\alias{fer_dataset}
\title{FER-2013 Facial Expression Dataset}
\usage{
fer_dataset(
  root = tempdir(),
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{(string, optional): Root directory for dataset storage,
the dataset will be stored under \code{root/fer2013}.}

\item{train}{(bool, optional): If True, creates dataset from
\code{training.pt}, otherwise from \code{test.pt}.}

\item{transform}{(callable, optional): A function/transform that  takes in an
PIL image and returns a transformed version. E.g,
\code{\link[=transform_random_crop]{transform_random_crop()}}.}

\item{target_transform}{(callable, optional): A function/transform that takes
in the target and transforms it.}

\item{download}{(bool, optional): If true, downloads the dataset from the
internet and puts it in root directory. If dataset is already downloaded,
it is not downloaded again.}
}
\value{
A torch dataset of class \code{fer_dataset}.
Each element is a named list:
\itemize{
\item \code{x}: a 48x48 grayscale array
\item \code{y}: an integer from 1 to 7 indicating the class index
}
}
\description{
Loads the FER-2013 dataset for facial expression recognition. The dataset contains grayscale images
(48x48) of human faces, each labeled with one of seven emotion categories:
\code{"Angry"}, \code{"Disgust"}, \code{"Fear"}, \code{"Happy"}, \code{"Sad"}, \code{"Surprise"}, and \code{"Neutral"}.
}
\details{
The dataset is split into:
\itemize{
\item \code{"Train"}: training images labeled as \code{"Training"} in the original CSV.
\item \code{"Test"}: includes both \code{"PublicTest"} and \code{"PrivateTest"} entries.
}
}
\examples{
\dontrun{
fer <- fer_dataset(train = TRUE, download = TRUE)
first_item <- fer[1]
first_item$x  # 48x48 grayscale array
first_item$y  # 4
fer$classes[first_item$y]  # "Happy"
}

}
\seealso{
Other datasets: 
\code{\link{coco_caption_dataset}()},
\code{\link{coco_detection_dataset}()},
\code{\link{fgvc_aircraft_dataset}()},
\code{\link{flowers102_dataset}()},
\code{\link{image_folder_dataset}()},
\code{\link{tiny_imagenet_dataset}()}
}
\concept{datasets}
