% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-fer.R
\name{fer_dataset}
\alias{fer_dataset}
\title{FER2013 dataset}
\usage{
fer_dataset(
  root = tempdir(),
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Character. Root directory where the \code{fer2013} folder exists or will be saved to if \code{download = TRUE}.}

\item{train}{Logical. If \code{TRUE}, loads the training split; if \code{FALSE}, loads the test split. Default is \code{TRUE}.}

\item{transform}{Optional function to transform input images.}

\item{target_transform}{Optional function to transform target labels.}

\item{download}{Logical. Whether to download the dataset if not found locally. Default is \code{FALSE}.}
}
\value{
A FER2013 dataset object.
}
\description{
FER2013 Dataset
}
\details{
Loads the FER2013 dataset for facial expression recognition, consisting of grayscale 48x48 images.
The dataset contains facial images categorized into seven emotion classes:
\itemize{
\item 0: Angry
\item 1: Disgust
\item 2: Fear
\item 3: Happy
\item 4: Sad
\item 5: Surprise
\item 6: Neutral
}
}
\examples{
\dontrun{
fer <- fer_dataset(train = TRUE, download = TRUE)
first_item <- fer[1]
# image in item 1
first_item$x
# label of item 1
first_item$y
# label name
first_item$class_name
}

}
