% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-flikr.R
\name{flickr8k_dataset}
\alias{flickr8k_dataset}
\title{Flickr8k Dataset}
\usage{
flickr8k_dataset(
  root,
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Character. Root directory where the dataset will be stored under \code{root/flickr8k}.}

\item{train}{Logical. If \code{TRUE}, loads the training split, otherwise the test split. Default is \code{TRUE}.}

\item{transform}{Optional function to transform input images after loading.}

\item{target_transform}{Optional function to transform captions.}

\item{download}{Logical. Whether to download the dataset if not found locally. Default is \code{FALSE}.}
}
\value{
A flickr8k_dataset object representing the dataset.
}
\description{
Flickr8k Dataset
}
\details{
Loads the Flickr8k dataset consisting of 8,000 images with five captions each.
The dataset is split into training and test sets with 6,000 and 1,000 images respectively.
Images are resized to a fixed size and captions are stored as lists of strings.
}
\examples{
\dontrun{
root_dir <- tempfile()
flickr8k <- flickr8k_dataset(root = root_dir, train = TRUE, download = TRUE)
first_item <- flickr8k[1]
# image tensor of first item
first_item$x
# list of captions of first item
first_item$y
}

}
