% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-flowers.R
\name{flowers102_dataset}
\alias{flowers102_dataset}
\title{Oxford Flowers 102 Dataset}
\usage{
flowers102_dataset(
  root = tempdir(),
  split = "train",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{(string, optional): Root directory for dataset storage,
the dataset will be stored under \code{root/flowers102}.}

\item{split}{(string, optional): One of \code{"train"}, \code{"val"}, or \code{"test"},
defines which subset of the data to load. Default is \code{"train"}.}

\item{transform}{(callable, optional): A function/transform that  takes in an
PIL image and returns a transformed version. E.g,
\code{\link[=transform_random_crop]{transform_random_crop()}}.}

\item{target_transform}{(callable, optional): A function/transform that takes
in the target and transforms it.}

\item{download}{(bool, optional): If true, downloads the dataset from the
internet and puts it in root directory. If dataset is already downloaded,
it is not downloaded again.}
}
\value{
An object of class \code{flowers102_dataset}, which behaves like a torch dataset.
Each element is a named list:
\itemize{
\item \code{x}: a 3 x H x W numeric array representing an RGB image.
\item \code{y}: an integer label indicating the class index.
}
}
\description{
Oxford Flowers 102 Dataset
}
\details{
Loads the Oxford 102 Category Flower Dataset. This dataset consists of 102 flower categories, with between 40 and 258 images per class.

The dataset is split into:
\itemize{
\item \code{"train"}: training subset with labels.
\item \code{"val"}: validation subset with labels.
\item \code{"test"}: test subset with labels (used for evaluation).
}
}
\examples{
\dontrun{
flowers <- flowers102_dataset(split = "train", download = TRUE)

# Define a custom collate function to resize images in the batch
resize_collate_fn <- function(batch) {
  xs <- lapply(batch, function(sample) {
    sample$x <- torch_tensor(sample$x)
    transform_resize(sample$x, c(224, 224))
  })
  xs <- torch_stack(xs)
  ys <- sapply(batch, function(sample) sample$y)
  list(x = xs, y = ys)
}

dl <- dataloader(dataset = flowers, batch_size = 4, collate_fn = resize_collate_fn)
batch <- dataloader_next(dataloader_make_iter(dl))
batch$x  # batched image tensors resized to 224x224
batch$y  # batched integer labels
}

}
