% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision_utils.R
\name{imagenet_classes}
\alias{imagenet_classes}
\title{Imagenet classes}
\usage{
imagenet_classes()
}
\description{
A list of imagenet classes that can be used to decode predictions from
imagenet pre-trained models.
}
\details{
The list has been adapted from a \href{https://gist.githubusercontent.com/yrevar/942d3a0ac09ec9e5eb3a/raw/238f720ff059c1f82f368259d1ca4ffa5dd8f9f5/imagenet1000_clsidx_to_labels.txt}{Gist}
For examples, see inception_v3 tests.
}
