% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagenet.R
\name{imagenet_classes}
\alias{imagenet_classes}
\alias{imagenet_label}
\title{ImageNet Class Labels}
\usage{
imagenet_classes()

imagenet_label(id)
}
\arguments{
\item{id}{Integer vector of 1-based class identifiers.}
}
\value{
A character vector with 1000 entries representing the ImageNet-1k
class labels.

A character vector with the labels associated with \code{id}.
}
\description{
Utilities for resolving ImageNet-1k class identifiers to their corresponding
human readable labels. The labels are retrieved from the same source used by
PyTorch's reference implementation.
}
\seealso{
Other class_resolution: 
\code{\link{coco_classes}()},
\code{\link{voc_classes}}

Other class_resolution: 
\code{\link{coco_classes}()},
\code{\link{voc_classes}}
}
\concept{class_resolution}
