% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transforms-geometry.R
\name{item_transform_pad}
\alias{item_transform_pad}
\title{Pad a dataset item}
\usage{
item_transform_pad(x, padding, fill = 0, padding_mode = "constant")
}
\arguments{
\item{x}{A dataset item, typically an \code{image_with_bounding_box} or
\code{image_with_segmentation_mask} object containing an image tensor
and associated target data.}

\item{padding}{(integer vector or integer): Padding on each border. If a
single integer is provided, this value is used to pad all borders. If an
integer vector of length 2 is provided, it gives the padding on
left/right and top/bottom respectively. If an integer vector of length 4
is provided, it gives the padding for the left, right, top and bottom
borders respectively.}

\item{fill}{(numeric vector or numeric): Pixel fill value for constant fill.
Default is \code{0}. If a numeric vector of length 3, it is used to fill R, G,
B channels respectively. This value is only used when the \code{padding_mode}
is constant. Only integer values are supported for Tensors.}

\item{padding_mode}{(character): Type of padding. Should be: \code{constant},
\code{edge}, \code{reflect} or \code{symmetric}. Default is \code{constant}. Mode \code{symmetric}
is not yet supported for Tensor inputs.
\itemize{
\item constant: pads with a constant value, this value is specified with fill
\item edge: pads with the last value on the edge of the image
\item reflect: pads with reflection of image (without repeating the last
value on the edge) padding \verb{[1, 2, 3, 4]} with 2 elements on both sides
in reflect mode will result in \verb{[3, 2, 1, 2, 3, 4, 3, 2]}
\item symmetric: pads with reflection of image (repeating the last value on
the edge) padding \verb{[1, 2, 3, 4]} with 2 elements on both sides in
symmetric mode will result in \verb{[2, 1, 1, 2, 3, 4, 4, 3]}
}}
}
\value{
A dataset item of the same class with the padded image and adjusted
target.
}
\description{
Pads the image inside a dataset item on all sides with the given "pad"
value. For detection items, bounding box coordinates are shifted to remain
correct after padding. For segmentation items, both the image and the masks
are padded.
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/b/b6/Felis_catus-cat_on_snow.jpg"
img <- base_loader(url) |> transform_to_tensor()

boxes <- torch_tensor(matrix(c(600, 200, 2880, 1860), ncol = 4), dtype = torch_float32())

before <- list(x = img, y = list(boxes = boxes, labels = "cat"))
class(before) <- c("image_with_bounding_box", "list")

after <- item_transform_pad(before, padding = 100)

before_plot <- draw_bounding_boxes(before, colors = "blue", width = 10)$to(torch_float())$div(255)
after_plot <- draw_bounding_boxes(after, colors = "red", width = 10)$to(torch_float())$div(255)

before_plot <- transform_resize(before_plot, c(2000, 3000))
after_plot <- transform_resize(after_plot, c(2000, 3000))

grid <- vision_make_grid(torch_stack(list(before_plot, after_plot)), scale = TRUE)
tensor_image_browse(grid)
}

}
\seealso{
Other item_unitary_transforms:
\code{\link[=item_transform_affine]{item_transform_affine()}},
\code{\link[=item_transform_center_crop]{item_transform_center_crop()}},
\code{\link[=item_transform_crop]{item_transform_crop()}},
\code{\link[=item_transform_hflip]{item_transform_hflip()}},
\code{\link[=item_transform_rotate]{item_transform_rotate()}},
\code{\link[=item_transform_vflip]{item_transform_vflip()}}
}
\concept{item_unitary_transforms}
