% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transforms-random-geometry.R
\name{item_transform_random_crop}
\alias{item_transform_random_crop}
\title{Randomly crop a dataset item}
\usage{
item_transform_random_crop(
  x,
  size,
  padding = NULL,
  pad_if_needed = FALSE,
  fill = 0,
  padding_mode = "constant"
)
}
\arguments{
\item{x}{A dataset item, typically an \code{image_with_bounding_box} or
\code{image_with_segmentation_mask} object containing an image tensor
and associated target data.}

\item{size}{Desired output size of the crop. If \code{size} is an
int instead of a sequence like \code{c(h, w)}, a square crop
\code{c(size, size)} is made. If a sequence of length 1 is provided it is
interpreted as \code{c(size[1], size[1])}.}

\item{padding}{(int or vector, optional) Optional padding on each border
of the image, applied before cropping. Default is \code{NULL}. If a single
int is provided this is used to pad all borders. If a vector of length 2 is
provided this is the padding on left/right and top/bottom respectively. If
a vector of length 4 is provided this is the padding for the left, top,
right and bottom borders respectively.}

\item{pad_if_needed}{(logical) It will pad the image if smaller than the
desired size to avoid raising an exception. Since cropping is done
after padding, the padding seems to be done at a random offset.}

\item{fill}{(number or tuple) Pixel fill value for constant fill. Default is
0. If a tuple of length 3, it is used to fill R, G, B channels
respectively. This value is only used when the padding_mode is constant.}

\item{padding_mode}{(str) Type of padding. Should be: constant, edge,
reflect or symmetric. Default is constant.}
}
\value{
A dataset item of the same class with the image and target randomly
cropped to the requested size.
}
\description{
Crops the image inside a dataset item at a random location. When a crop
occurs, the same coordinate adjustments used by
\code{\link{item_transform_crop}} are applied to bounding boxes and masks so
that the targets stay aligned with the cropped image.
}
\details{
If the image is smaller than the requested crop size, it can optionally be
padded first (see \code{padding} and \code{pad_if_needed}).
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/b/b6/Felis_catus-cat_on_snow.jpg"
img <- base_loader(url) |> transform_to_tensor()

boxes <- torch_tensor(matrix(c(600, 200, 2880, 1860), ncol = 4), dtype = torch_float32())

before <- list(x = img, y = list(boxes = boxes, labels = {"CAT"}))
class(before) <- c("image_with_bounding_box", "list")

after <- item_transform_random_crop(before, size = c(800, 1200))

before_plot <- draw_bounding_boxes(before, colors = "blue", width = 10)$to(torch_float())$div(255)
after_plot <- draw_bounding_boxes(after, colors = "red", width = 10)$to(torch_float())$div(255)

# the crop changes the image size, so resize before stacking into the grid
grid <- vision_make_grid(
  torch_stack(list(transform_resize(before_plot, c(600, 600)),
                   transform_resize(after_plot, c(600, 600)))),
  scale = TRUE
)
tensor_image_browse(grid)
}

}
\seealso{
Other item_random_transforms:
\code{\link[=item_transform_random_affine]{item_transform_random_affine()}},
\code{\link[=item_transform_random_horizontal_flip]{item_transform_random_horizontal_flip()}},
\code{\link[=item_transform_random_resize_crop]{item_transform_random_resize_crop()}},
\code{\link[=item_transform_random_rotation]{item_transform_random_rotation()}},
\code{\link[=item_transform_random_vertical_flip]{item_transform_random_vertical_flip()}}
}
\concept{item_random_transforms}
