% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transforms-geometry.R
\name{item_transform_random_erasing}
\alias{item_transform_random_erasing}
\title{Randomly erase a rectangular region of a dataset item}
\usage{
item_transform_random_erasing(
  x,
  p = 0.5,
  scale = c(0.02, 0.33),
  ratio = c(0.3, 3.3),
  value = 0,
  inplace = FALSE
)
}
\arguments{
\item{x}{A dataset item, typically an \code{image_with_bounding_box} or
\code{image_with_segmentation_mask} object containing an image tensor
and associated target data.}

\item{p}{(numeric): Probability that the random erasing operation will be
performed. Default is 0.5.}

\item{scale}{(numeric vector of length 2): Range of proportion of erased
area against input image.}

\item{ratio}{(numeric vector of length 2): Range of aspect ratio of erased
area.}

\item{value}{(numeric vector or numeric or character): Erasing value.
Default is \code{0}. If a single numeric value, it is used to erase all
pixels. If a numeric vector of length 3, it is used to erase R, G, B
channels respectively. If the string \code{"random"}, erasing each pixel with
random values.}

\item{inplace}{(logical): Boolean to make this transform inplace. Default
set to \code{FALSE}.}
}
\value{
A dataset item of the same class. With probability \code{p}, the
image is randomly erased; otherwise it is returned unchanged.
}
\description{
Randomly selects a rectangular region in the image inside a dataset item and
erases its pixel values with a given probability. Only the image pixels are
modified: bounding boxes and masks are left unchanged.
}
\details{
'Random Erasing Data Augmentation' by Zhong \emph{et al.} See
\url{https://arxiv.org/abs/1708.04896}
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/b/b6/Felis_catus-cat_on_snow.jpg"
img <- base_loader(url) |> transform_to_tensor()

boxes <- torch_tensor(matrix(c(600, 200, 2880, 1860), ncol = 4), dtype = torch_float32())

before <- list(x = img, y = list(boxes = boxes, labels = "cat"))
class(before) <- c("image_with_bounding_box", "list")

after <- item_transform_random_erasing(before)

before_plot <- draw_bounding_boxes(before, colors = "blue", width = 10)$to(torch_float())$div(255)
after_plot <- draw_bounding_boxes(after, colors = "red", width = 10)$to(torch_float())$div(255)

grid <- vision_make_grid(torch_stack(list(before_plot, after_plot)), scale = TRUE)
tensor_image_browse(grid)
}

}
\seealso{
Other item_random_transforms:
\code{\link[=item_transform_random_horizontal_flip]{item_transform_random_horizontal_flip()}},
\code{\link[=item_transform_random_vertical_flip]{item_transform_random_vertical_flip()}}
}
\concept{item_random_transforms}
