% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transforms-random-geometry.R
\name{item_transform_random_horizontal_flip}
\alias{item_transform_random_horizontal_flip}
\title{Randomly horizontally flip a dataset item}
\usage{
item_transform_random_horizontal_flip(x, p = 0.5)
}
\arguments{
\item{x}{A dataset item, typically an \code{image_with_bounding_box} or
\code{image_with_segmentation_mask} object containing an image tensor
and associated target data.}

\item{p}{Probability of the item being flipped. Default is 0.5.}
}
\value{
A dataset item of the same class. With probability \code{p}, the
image and targets are horizontally flipped; otherwise they are returned
unchanged.
}
\description{
Randomly flips the image inside a dataset item horizontally with a given
probability. When a flip occurs, the same coordinate adjustments used by
\code{\link{item_transform_hflip}} are applied to bounding boxes and masks.
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/b/b6/Felis_catus-cat_on_snow.jpg"
img <- base_loader(url) |> transform_to_tensor()

boxes <- torch_tensor(matrix(c(600, 200, 2880, 1860), ncol = 4), dtype = torch_float32())

before <- list(x = img, y = list(boxes = boxes, labels = {"CAT"}))
class(before) <- c("image_with_bounding_box", "list")

after <- item_transform_random_horizontal_flip(before)

before_plot <- draw_bounding_boxes(before, colors = "blue", width = 10)$to(torch_float())$div(255)
after_plot <- draw_bounding_boxes(after, colors = "red", width = 10)$to(torch_float())$div(255)

grid <- vision_make_grid(torch_stack(list(before_plot, after_plot)), scale = TRUE)
tensor_image_browse(grid)
}

}
\seealso{
Other item_random_transforms:
\code{\link[=item_transform_random_affine]{item_transform_random_affine()}},
\code{\link[=item_transform_random_crop]{item_transform_random_crop()}},
\code{\link[=item_transform_random_resize_crop]{item_transform_random_resize_crop()}},
\code{\link[=item_transform_random_vertical_flip]{item_transform_random_vertical_flip()}}
}
\concept{item_random_transforms}
