% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transforms-geometry.R
\name{item_transform_rotate}
\alias{item_transform_rotate}
\title{Rotate dataset item}
\usage{
item_transform_rotate(x, angle, interpolation = 2, expand = TRUE, fill = 0)
}
\arguments{
\item{x}{A dataset item, typically an \code{image_with_bounding_box} object
containing an image tensor and associated target data (boxes, labels).}

\item{angle}{(numeric): Rotation angle value in degrees, counter-clockwise.}

\item{interpolation}{(integer, optional): Interpolation mode. \code{0} for
nearest, \code{2} for bilinear. Default is \code{0} (nearest).}

\item{expand}{(logical, optional): Optional expansion flag. If \code{TRUE},
expands the output to make it large enough to hold the entire rotated
image. If \code{FALSE} or omitted, make the output image the same size as the
input image. Note that the expand flag assumes rotation around the center
and no translation.}

\item{fill}{(numeric vector or numeric): Pixel fill value for area outside
the rotated image. If a single numeric value, the value is used for all
channels. Defaults to \code{0} for all channels. This option is not supported for Tensor input. Fill
value for the area outside the transform in the output image is always 0.}
}
\value{
An \code{image_with_rotated_box} object with the rotated image and
converted boxes in xyxyr format.
}
\description{
Rotates a dataset item by a given angle around its center.
The canvas is expanded so that the entire rotated image is visible with no
cropping. Empty regions are filled with given fill color.
}
\details{
The bounding boxes (if present) are shifted to account for the expanded
canvas and converted to rotated format via
\code{\link{target_transform_rotate}}.
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/6/66/The_Leaning_Tower_of_Pisa_SB.jpeg"

img <- base_loader(url) |>
  transform_to_tensor()

boxes <- torch_tensor(matrix(c(720, 620, 1900, 3700), ncol = 4), dtype = torch_float32())

before <- list(x = img, y = list(boxes = boxes, labels = {"Leaning Tower of Pisa"}))
class(before) <- c("image_with_bounding_box", "list")

after <- item_transform_rotate(before, angle = 30)

before_plot <- draw_bounding_boxes(before, colors = "blue", width = 10)
after_plot <- draw_bounding_boxes(after, colors = "red", width = 10)
tensor_image_browse(before_plot)
tensor_image_browse(after_plot)
}

}
\concept{item_unitary_transforms}
