% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-transform-geometry.R
\name{item_transform_rotate}
\alias{item_transform_rotate}
\title{Rotate image}
\usage{
item_transform_rotate(x, angle = 0)
}
\arguments{
\item{x}{A torch tensor of shape \code{(C, H, W)}, or an
\code{image_with_bounding_box} object.}

\item{angle}{(numeric): Rotation angle in degrees (counter-clockwise).
Default is \code{0}.}
}
\value{
The rotated image tensor with expanded resolution, or an
\code{image_with_rotated_box} object when the input is
\code{image_with_bounding_box}.
}
\description{
Rotates an image by a given angle around its center. The canvas is expanded
so that the entire rotated image is visible with no cropping. Empty regions
are filled with black.
}
\details{
When passed an \code{image_with_bounding_box} object, the bounding boxes are
also rotated via an internal call to \code{\link{target_transform_rotate_box}}.
}
\examples{
\dontrun{
url <- "https://upload.wikimedia.org/wikipedia/commons/c/c6/Jumping_dog.JPG"

img <- base_loader(url) |>
  transform_to_tensor() |>
  transform_resize(c(300, 500))

before_plot <- draw_bounding_boxes(img, colors = "blue", width = 4)
after_plot <- draw_bounding_boxes(item_transform_rotate(img, angle = 30), colors = "red", width = 4)

grid <- vision_make_grid(torch_stack(list(before_plot, after_plot))$to(torch_float32()), scale = TRUE)
tensor_image_browse(grid)
}

}
\seealso{
Other item_unitary_transforms:
\code{\link[=target_transform_rotate_box]{target_transform_rotate_box()}}
}
\concept{item_unitary_transforms}
