% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-lfw.R
\name{lfw_dataset}
\alias{lfw_dataset}
\alias{lfw_people_dataset}
\alias{lfw_pairs_dataset}
\title{LFW Datasets}
\usage{
lfw_people_dataset(
  root = tempdir(),
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)

lfw_pairs_dataset(
  root = tempdir(),
  train = TRUE,
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{root}{Root directory for dataset storage. The dataset will be stored under \code{root/lfw_people} or \code{root/lfw_pairs}.}

\item{train}{Logical. If TRUE, use the training set; otherwise, use the test set. Not applicable to all datasets.}

\item{transform}{Optional. A function that takes an image and returns a transformed version (e.g., normalization, cropping).}

\item{target_transform}{Optional. A function that transforms the label.}

\item{download}{Logical. If TRUE, downloads the dataset to \verb{root/}. If the dataset is already present, download is skipped.}
}
\value{
A torch dataset object \code{lfw_people_dataset} or \code{lfw_pairs_dataset}.
Each element is a named list with:
\itemize{
\item \code{x}:
\itemize{
\item For \code{lfw_people_dataset}: a H x W x 3 numeric array representing a single RGB image.
\item For \code{lfw_pairs_dataset}: a list of two H x W x 3 numeric arrays representing a pair of RGB images.
}
\item \code{y}:
\itemize{
\item For \code{lfw_people_dataset}: an integer index from 1 to the number of identities in the dataset.
\item For \code{lfw_pairs_dataset}: 1 if the pair shows the same person, 2 if different people.
}
}
}
\description{
Labelled Faces in the Wild (LFW) Datasets
}
\details{
The LFW dataset collection provides facial images for evaluating face recognition systems.
It includes two variants:
\itemize{
\item \code{lfw_people_dataset}: A \strong{multi-class classification} dataset where each image is labelled by person identity.
\item \code{lfw_pairs_dataset}: A \strong{face verification} dataset containing image pairs with binary labels (same or different person).
}

The LFW People dataset uses the "deep funneled" version of LFW, offering aligned and cropped RGB face images.
\itemize{
\item Training split: 9525 images, 4038 identities
\item Test split: 3708 images, 1711 identities
The dataset is downloaded from \href{https://huggingface.co/datasets/JimmyUnleashed/LFW}{Hugging Face}.
}
}
\examples{
\dontrun{
# Load training data for LFW People Dataset
lfw <- lfw_people_dataset(download = TRUE, train = TRUE)
first_item <- lfw[1]
first_item$x  # RGB image
first_item$y  # Label index
lfw$classes[first_item$y]  # person's name (e.g., "AJ_Cook")

# Load test data for LFW People Dataset
lfw_test <- lfw_people_dataset(download = TRUE, train = FALSE)
first_item <- lfw_test[1]
first_item$x  # RGB image
first_item$y  # Label index
lfw_test$classes[first_item$y]  # e.g., "AJ_Lamas"

# Load training data for LFW Pairs Dataset
lfw <- lfw_pairs_dataset(download = TRUE, train = TRUE)
first_item <- lfw[1]
first_item$x  # List of 2 RGB Images
first_item$x[[1]]  # RGB Image
first_item$x[[2]]  # RGB Image
first_item$y  # Label index
lfw$classes[first_item$y]  # Class Name

# Load test data for LFW Pairs Dataset
lfw <- lfw_pairs_dataset(download = TRUE, train = FALSE)
first_item <- lfw[1]
first_item$x  # List of 2 RGB Images
first_item$x[[1]]  # RGB Image
first_item$x[[2]]  # RGB Image
first_item$y  # Label index
lfw$classes[first_item$y]  # Class Name
}

}
\seealso{
Other classification_dataset: 
\code{\link{caltech_dataset}},
\code{\link{cifar10_dataset}()},
\code{\link{eurosat_dataset}()},
\code{\link{fer_dataset}()},
\code{\link{fgvc_aircraft_dataset}()},
\code{\link{flowers102_dataset}()},
\code{\link{mnist_dataset}()},
\code{\link{oxfordiiitpet_dataset}()},
\code{\link{tiny_imagenet_dataset}()}
}
\concept{classification_dataset}
