% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder-dataset.R
\name{magick_loader}
\alias{magick_loader}
\title{Loads an image from a path using imagemagick}
\usage{
magick_loader(path)
}
\arguments{
\item{path}{path to the image to load from.}
}
\description{
Loads an image from a path using imagemagick
}
