% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder-dataset.R
\name{magick_loader}
\alias{magick_loader}
\title{Load an Image using ImageMagick}
\usage{
magick_loader(path)
}
\arguments{
\item{path}{path or URL to load the image from.}
}
\value{
an magick-image object as result of \code{image_read()}
}
\description{
Load an image located at \code{path} using the \code{{magick}} package.
}
