% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-alexnet.R
\name{model_alexnet}
\alias{model_alexnet}
\title{AlexNet Model Architecture}
\usage{
model_alexnet(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{(bool): If TRUE, returns a model pre-trained on ImageNet.}

\item{progress}{(bool): If TRUE, displays a progress bar of the download to
stderr.}

\item{...}{other parameters passed to the model intializer. currently only
\code{num_classes} is used.}
}
\description{
AlexNet model architecture from the
\href{https://arxiv.org/abs/1404.5997}{One weird trick...} paper.
}
\seealso{
Other classification_model: 
\code{\link{model_convnext}},
\code{\link{model_efficientnet}},
\code{\link{model_efficientnet_v2}},
\code{\link{model_inception_v3}()},
\code{\link{model_maxvit}()},
\code{\link{model_mobilenet_v2}()},
\code{\link{model_mobilenet_v3}},
\code{\link{model_resnet}},
\code{\link{model_vgg}},
\code{\link{model_vit}}
}
\concept{classification_model}
