% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-efficientnet.R
\name{model_efficientnet}
\alias{model_efficientnet}
\alias{model_efficientnet_b0}
\alias{model_efficientnet_b1}
\alias{model_efficientnet_b2}
\alias{model_efficientnet_b3}
\alias{model_efficientnet_b4}
\alias{model_efficientnet_b5}
\alias{model_efficientnet_b6}
\alias{model_efficientnet_b7}
\title{EfficientNet implementation}
\usage{
model_efficientnet_b0(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_b1(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_b2(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_b3(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_b4(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_b5(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_b6(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_b7(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{(bool): If TRUE, returns a model pre-trained on ImageNet.}

\item{progress}{(bool): If TRUE, displays a progress bar of the download to
stderr.}

\item{...}{Other parameters passed to the EfficientNet implementation.}
}
\description{
EfficientNet models are based on the architecture described in
\href{https://arxiv.org/abs/1905.11946}{EfficientNet: Rethinking Model Scaling for Convolutional Neural Networks}.
}
\section{Functions}{
\itemize{
\item \code{model_efficientnet_b0()}: EfficientNet B0 model

\item \code{model_efficientnet_b1()}: EfficientNet B1 model

\item \code{model_efficientnet_b2()}: EfficientNet B2 model

\item \code{model_efficientnet_b3()}: EfficientNet B3 model

\item \code{model_efficientnet_b4()}: EfficientNet B4 model

\item \code{model_efficientnet_b5()}: EfficientNet B5 model

\item \code{model_efficientnet_b6()}: EfficientNet B6 model

\item \code{model_efficientnet_b7()}: EfficientNet B7 model

}}
\seealso{
Other models: 
\code{\link{model_alexnet}()},
\code{\link{model_inception_v3}()},
\code{\link{model_mobilenet_v2}()},
\code{\link{model_resnet}},
\code{\link{model_vgg}}
}
\concept{models}
