% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-efficientnetv2.R
\name{model_efficientnet_v2}
\alias{model_efficientnet_v2}
\alias{model_efficientnet_v2_s}
\alias{model_efficientnet_v2_m}
\alias{model_efficientnet_v2_l}
\title{EfficientNetV2 Models}
\usage{
model_efficientnet_v2_s(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_v2_m(pretrained = FALSE, progress = TRUE, ...)

model_efficientnet_v2_l(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{(bool): If TRUE, returns a model pre-trained on ImageNet.}

\item{progress}{(bool): If TRUE, displays a progress bar of the download to
stderr.}

\item{...}{Other parameters passed to the model implementation, such as
\code{num_classes} to change the output dimension.}
}
\description{
Constructs EfficientNetV2 model architectures as described in
\emph{EfficientNetV2: Smaller Models and Faster Training}.
}
\section{Functions}{
\itemize{
\item \code{model_efficientnet_v2_s()}: EfficientNetV2-S model

\item \code{model_efficientnet_v2_m()}: EfficientNetV2-M model

\item \code{model_efficientnet_v2_l()}: EfficientNetV2-L model

}}
\examples{
\dontrun{
model <- model_efficientnet_v2_s()
input <- torch::torch_randn(1, 3, 224, 224)
output <- model(input)
}
}
\seealso{
\code{\link{model_efficientnet}}

Other models: 
\code{\link{model_alexnet}()},
\code{\link{model_efficientnet}},
\code{\link{model_inception_v3}()},
\code{\link{model_mobilenet_v2}()},
\code{\link{model_resnet}},
\code{\link{model_vgg}}
}
\concept{models}
