% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-maxvit.R
\name{model_maxvit}
\alias{model_maxvit}
\title{MaxViT Model}
\usage{
model_maxvit(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{(bool): If TRUE, the pre-trained weights are downloaded.
Currently only a pre-trained TorchScript model is available and this
argument is ignored.}

\item{progress}{(bool): If TRUE, displays a progress bar of the download to
stderr.}

\item{...}{Not used.}
}
\description{
Loads a pre-trained MaxViT classification model as described in
\href{https://arxiv.org/abs/2204.01697}{MaxViT: Multi-Axis Vision Transformer}.
The model expects input tensors of shape \code{(batch_size, 3, 224, 224)}
and returns classification logits with \code{1000} output classes.
}
