% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-resnet.R
\name{model_resnet18}
\alias{model_resnet18}
\title{ResNet-18 Model Architecture}
\usage{
model_resnet18(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{(bool): If TRUE, returns a model pre-trained on ImageNet.}

\item{progress}{(bool): If TRUE, displays a progress bar of the download to
stderr.}

\item{...}{Other parameters passed to the resnet model.}
}
\description{
ResNet-18 model architecture from
\href{https://arxiv.org/pdf/1512.03385.pdf}{Deep Residual Learning for Image Recognition}
}
\seealso{
Other models: 
\code{\link{model_alexnet}()},
\code{\link{model_mobilenet_v2}()}
}
\concept{models}
