% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_convnext.R
\name{model_resnext}
\alias{model_resnext}
\title{ConvNeXt Implementation}
\arguments{
\item{...}{Other parameters passed to the model implementation.}
}
\description{
Implements the ConvNeXt architecture from \href{https://arxiv.org/pdf/2201.03545}{ConvNeXt: A ConvNet for the 2020s}
}
\seealso{
Other classification_model: 
\code{\link{model_alexnet}()},
\code{\link{model_efficientnet}},
\code{\link{model_efficientnet_v2}},
\code{\link{model_inception_v3}()},
\code{\link{model_maxvit}()},
\code{\link{model_mobilenet_v2}()},
\code{\link{model_mobilenet_v3}},
\code{\link{model_resnet}},
\code{\link{model_vgg}},
\code{\link{model_vit}}
}
\concept{classification_model}
