% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-vit.R
\name{model_vit}
\alias{model_vit}
\alias{model_vit_b_16}
\alias{model_vit_b_32}
\alias{model_vit_l_16}
\alias{model_vit_l_32}
\alias{model_vit_h_14}
\title{Vision Transformer Implementation}
\usage{
model_vit_b_16(pretrained = FALSE, progress = TRUE, ...)

model_vit_b_32(pretrained = FALSE, progress = TRUE, ...)

model_vit_l_16(pretrained = FALSE, progress = TRUE, ...)

model_vit_l_32(pretrained = FALSE, progress = TRUE, ...)

model_vit_h_14(pretrained = FALSE, progress = TRUE, ...)
}
\arguments{
\item{pretrained}{Whether to load TorchScript weights.}

\item{progress}{Show progress bar during download.}

\item{...}{Additional arguments passed to the model constructor.}
}
\description{
Vision Transformer models implementation on \href{https://arxiv.org/abs/2010.11929}{An Image is Worth 16x16 Words}
}
\section{Functions}{
\itemize{
\item \code{model_vit_b_16()}: ViT-B/16 model (Base, 16×16 patch size)

\item \code{model_vit_b_32()}: ViT-B/32 model (Base, 32×32 patch size)

\item \code{model_vit_l_16()}: ViT-L/16 model (Base, 16×16 patch size)

\item \code{model_vit_l_32()}: ViT-L/32 model (Base, 32×32 patch size)

\item \code{model_vit_h_14()}: ViT-H/14 model (Base, 14×14 patch size)

}}
\seealso{
Other models: 
\code{\link{model_alexnet}()},
\code{\link{model_inception_v3}()},
\code{\link{model_mobilenet_v2}()},
\code{\link{model_resnet}},
\code{\link{model_vgg}}
}
\concept{models}
