% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms-coco-captions-dataset.R
\name{ms_coco_captions_dataset}
\alias{ms_coco_captions_dataset}
\title{MS COCO Captions Dataset}
\usage{
ms_coco_captions_dataset(root, split = "val", download = FALSE, ...)
}
\arguments{
\item{root}{Path to the root directory for data storage.}

\item{split}{"train" or "val".}

\item{download}{Whether to download the dataset if not present.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Loads the MS COCO Captions dataset.
}
