% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places365-dataset.R
\name{places365_dataset}
\alias{places365_dataset}
\title{Places365 dataset}
\usage{
places365_dataset(root, split = "train", download = FALSE, ...)
}
\arguments{
\item{root}{directory path to download the dataset.}

\item{split}{dataset split, \code{train}, \code{val}, or \code{test}.}

\item{download}{whether to download the dataset.}

\item{...}{other arguments passed to \code{\link[=image_folder_dataset]{image_folder_dataset()}}.}
}
\description{
Prepares the Places365 dataset and optionally downloads it.
}
\seealso{
Other dataset: 
\code{\link{tiny_imagenet_dataset}()}
}
\concept{dataset}
