% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-boxes.R
\name{remove_small_boxes}
\alias{remove_small_boxes}
\title{Remove Small Boxes}
\usage{
remove_small_boxes(boxes, min_size)
}
\arguments{
\item{boxes}{(Tensor[N, 4]): boxes in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format
with
\itemize{
\item \eqn{0 \leq x_{min} < x_{max}} and
\item \eqn{0 \leq y_{min} < y_{max}}.
}}

\item{min_size}{(float): minimum size}
}
\value{
keep (Tensor[K]): indices of the boxes that have both sides
larger than min_size
}
\description{
Remove boxes which contains at least one side smaller than min_size.
}
