% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection-rf100-biology.R
\name{rf100_biology_collection}
\alias{rf100_biology_collection}
\title{RoboFlow 100  Biology dataset Collection}
\usage{
rf100_biology_collection(
  dataset,
  split = c("train", "test", "valid"),
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{dataset}{Dataset to select within \code{c("stomata_cell", "blood_cell", "parasite", "cell",
"liquid_crystals", "bacteria", "cotton_desease","mitosis", "phage", "liver_desease", "asbestos")}.}

\item{split}{the subset of the dataset to choose between \code{c("train", "test", "valid")}.}

\item{transform}{Optional transform function applied to the image.}

\item{target_transform}{Optional transform function applied to the target.}

\item{download}{Logical. If TRUE, downloads the dataset if not present at \code{root}.}
}
\value{
A torch dataset. Each element is a named list with:
\itemize{
\item \code{x}: H x W x 3 array representing the image.
\item \code{y}: a list containing the target with:
\itemize{
\item \code{labels}: character vector of object class names.
\item \code{boxes}: a tensor of shape (N, 4) with bounding boxes, if any, in \eqn{(x_{min}, y_{min}, x_{max}, y_{max})} format.
}
}

The returned item inherits the class \code{image_with_bounding_box} so it can be
visualised with helper functions such as \code{\link[=draw_bounding_boxes]{draw_bounding_boxes()}}.
}
\description{
Loads one of the \href{https://universe.roboflow.com/browse/biology}{RoboFlow 100 Biology} datasets with COCO-style
bounding box annotations for object detection tasks.
}
\examples{
\dontrun{
ds <- rf100_biology_collection(
  dataset = "stomata_cell",
  split = "test",
  transform = transform_to_tensor,
  download = TRUE
)
item <- ds[1]
boxed <- draw_bounding_boxes(item)
tensor_image_browse(boxed)
}

}
\seealso{
Other detection_dataset: 
\code{\link{coco_detection_dataset}()},
\code{\link{pascal_voc_datasets}},
\code{\link{rf100_document_collection}()},
\code{\link{rf100_medical_collection}()},
\code{\link{rf100_underwater_collection}()}
}
\concept{detection_dataset}
