% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-segmentation.R
\name{target_transform_coco_masks}
\alias{target_transform_coco_masks}
\title{Target Transform: COCO Polygon Segmentation to Masks}
\usage{
target_transform_coco_masks(y)
}
\arguments{
\item{y}{List containing COCO target with \code{segmentation}, \code{image_height}, \code{image_width}}
}
\value{
Modified y list with added \code{masks} field (N, H, W) boolean tensor
}
\description{
Converts COCO-style polygon segmentation annotations to boolean mask tensors.
Use as \code{target_transform} in \code{coco_detection_dataset()}.
}
\examples{
\dontrun{
ds <- coco_detection_dataset(
  root = "data",
  target_transform = target_transform_coco_masks
)
item <- ds[1]
draw_segmentation_masks(item)
}

}
