% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-generics.R
\name{transform_adjust_gamma}
\alias{transform_adjust_gamma}
\title{Adjust the gamma of an RGB image}
\usage{
transform_adjust_gamma(img, gamma, gain = 1)
}
\arguments{
\item{img}{A \code{magick-image}, \code{array} or \code{torch_tensor}.}

\item{gamma}{(float): Non negative real number, same as \eqn{\gamma} in the
equation. gamma larger than 1 make the shadows darker, while gamma smaller
than 1 make dark regions lighter.}

\item{gain}{(float): The constant multiplier.}
}
\description{
Also known as Power Law Transform. Intensities in RGB mode are adjusted
based on the following equation:
\deqn{
    I_{\mbox{out}} = 255 \times \mbox{gain} \times \left
    (\frac{I_{\mbox{in}}}{255}\right)^{\gamma}
    }
}
\details{
See \href{https://en.wikipedia.org/wiki/Gamma_correction}{Gamma Correction} for more details.
}
\seealso{
Other unitary_transforms: 
\code{\link{transform_adjust_brightness}()},
\code{\link{transform_adjust_contrast}()},
\code{\link{transform_adjust_hue}()},
\code{\link{transform_adjust_saturation}()},
\code{\link{transform_affine}()},
\code{\link{transform_center_crop}()},
\code{\link{transform_convert_image_dtype}()},
\code{\link{transform_crop}()},
\code{\link{transform_grayscale}()},
\code{\link{transform_hflip}()},
\code{\link{transform_linear_transformation}()},
\code{\link{transform_normalize}()},
\code{\link{transform_pad}()},
\code{\link{transform_perspective}()},
\code{\link{transform_resize}()},
\code{\link{transform_rgb_to_grayscale}()},
\code{\link{transform_rotate}()},
\code{\link{transform_to_tensor}()},
\code{\link{transform_vflip}()}
}
\concept{unitary_transforms}
