% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_adjust_saturation}
\alias{transform_adjust_saturation}
\title{Adjust color saturation of an image.}
\usage{
transform_adjust_saturation(img, saturation_factor)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be adjusted.}

\item{saturation_factor}{(float):  How much to adjust the saturation. 0 will
give a black and white image, 1 will give the original image while
2 will enhance the saturation by a factor of 2.}
}
\value{
Magick Image or Tensor: Saturation adjusted image.
}
\description{
Adjust color saturation of an image.
}
