% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_center_crop}
\alias{transform_center_crop}
\title{Crops the given image at the center.}
\usage{
transform_center_crop(img, output_size)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be cropped.}

\item{output_size}{(sequence or int): (height, width) of the crop box. If int
or sequence with single int it is used for both directions.}
}
\value{
Magick Image or Tensor: Cropped image.
}
\description{
The image can be a PIL Image or a Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading
dimensions
}
