% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_convert_image_dtype}
\alias{transform_convert_image_dtype}
\title{Convert a tensor image to the given \code{dtype} and scale the values accordingly}
\usage{
transform_convert_image_dtype(image, dtype = torch::torch_float())
}
\arguments{
\item{image}{(torch.Tensor): Image to be converted}

\item{dtype}{(torch.dtype): Desired data type of the output}
}
\value{
(torch.Tensor): Converted image
}
\description{
Convert a tensor image to the given \code{dtype} and scale the values accordingly
}
\note{
When converting from a smaller to a larger integer \code{dtype} the maximum values
are \strong{not} mapped exactly. If converted back and forth, this mismatch has no
effect.
}
\section{Raises}{


runtime_error: When trying to cast \code{torch_float32} to \code{torch_int32} or
\code{torch_int64} as well as for trying to cast \code{torch_float64} to \code{torch_int64}.
These conversions might lead to overflow errors since the floating point
\code{dtype} cannot store consecutive integers over the whole range of the integer
\code{dtype}.
}

