% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_resize}
\alias{transform_resize}
\title{Resize the input image to the given size.
The image can be a Magick Image or a torch Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading dimensions}
\usage{
transform_resize(img, size, interpolation = 2)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be resized.}

\item{size}{(sequence or int): Desired output size. If size is a sequence
like (h, w), the output size will be matched to this. If size is an int,
the smaller edge of the image will be matched to this number maintaining
the aspect ratio. i.e, if height > width, then image will be rescaled to
\eqn{\left(\text{size} \times \frac{\text{height}}{\text{width}}, \text{size}\right)}.}

\item{interpolation}{(int, optional): Desired interpolation enum defined by \code{filters}.
Default is \code{2 = BILINEAR}. If input is Tensor, only \code{0 = NEAREST}, \code{2 = BILINEAR}
and \code{3 = BICUBIC} are supported.}
}
\value{
Magick Image or Tensor: Resized image.
}
\description{
Resize the input image to the given size.
The image can be a Magick Image or a torch Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading dimensions
}
