% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_resized_crop}
\alias{transform_resized_crop}
\title{Crop the given image and resize it to desired size.}
\usage{
transform_resized_crop(img, top, left, height, width, size, interpolation = 2)
}
\arguments{
\item{img}{(Magick Image or Tensor): Image to be cropped. (1,1) denotes the
top left corner of the image.}

\item{top}{(int): Vertical component of the top left corner of the crop box.}

\item{left}{(int): Horizontal component of the top left corner of the crop box.}

\item{height}{(int): Height of the crop box.}

\item{width}{(int): Width of the crop box.}

\item{size}{(sequence or int): Desired output size. If size is a sequence
like (h, w), the output size will be matched to this. If size is an int,
the smaller edge of the image will be matched to this number maintaining
the aspect ratio. i.e, if height > width, then image will be rescaled to
\eqn{\left(\text{size} \times \frac{\text{height}}{\text{width}}, \text{size}\right)}.}

\item{interpolation}{(int, optional): Desired interpolation enum defined by \code{filters}.
Default is \code{2 = BILINEAR}. If input is Tensor, only \code{0 = NEAREST}, \code{2 = BILINEAR}
and \code{3 = BICUBIC} are supported.}
}
\value{
Magick Image or Tensor: Cropped image.
}
\description{
The image can be a Magick Image or a Tensor, in which case it is expected
to have \verb{[..., H, W]} shape, where ... means an arbitrary number of leading
dimensions
}
