% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-segmentation.R
\name{transform_sahi_crop}
\alias{transform_sahi_crop}
\title{Transform: SAHI Image Slicing}
\usage{
transform_sahi_crop(x, size = c(512L, 512L), overlap_size_ratio = c(0.2, 0.2))
}
\arguments{
\item{x}{Image input. Can be a \code{torch_tensor} of shape \verb{(C, H, W)} or a
supported image type (e.g., \code{magick-image}, \code{array}). Non-tensor inputs
are converted using \code{transform_to_tensor()} before slicing.}

\item{size}{Integer vector of length 2 containing crop height and width
in the form \code{c(height, width)}.}

\item{overlap_size_ratio}{Numeric vector of length 2 containing vertical
and horizontal overlap ratios in the form
\code{c(overlap_height_ratio, overlap_width_ratio)}.}
}
\value{
A list with:
\describe{
\item{images}{List of cropped image tensors.}
\item{crop_windows}{List of crop windows. Each crop window contains
\code{top}, \code{left}, \code{height} and \code{width} fields.}
}
}
\description{
Splits a large image into overlapping crops following the
SAHI (Slicing Aided Hyper Inference) approach. This transform is useful
for object detection workflows where downscaling large images would
otherwise remove small object details.
}
\details{
The transform returns both the cropped image tiles and the crop window
metadata required to later transform detection targets using
\code{target_transform_sahi_crop()}.

Use as \code{transform} in image datasets.
}
\note{
If \code{x} is not a \code{torch_tensor}, \code{transform_to_tensor()} is applied.
}
\examples{
\dontrun{
img_url <- "https://raw.githubusercontent.com/obss/sahi/main/demo/demo_data/small-vehicles1.jpeg"

img <- base_loader(img_url) \%>\%
  transform_to_tensor()

item <- transform_sahi_crop(img)

item$crop_windows[[1]]$top      # Top coordinate of the first crop in the original image
item$crop_windows[[1]]$left     # Left coordinate of the first crop in the original image
item$crop_windows[[1]]$height   # Height of the first crop
item$crop_windows[[1]]$width    # Width of the first crop

item$images[[1]]                # First cropped image tensor

length(item$images)             # Number of generated image crops

crops <- torch_stack(
  item$images[1:length(item$images)]
)

grid <- vision_make_grid(
  crops,
  scale = TRUE,
  num_rows = 3
)

tensor_image_browse(grid)       # Displays a grid of the generated crops
}

}
\references{
SAHI Documentation:
https://obss.github.io/sahi/slicing/
}
