% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms-generics.R
\name{transform_sahi_crop}
\alias{transform_sahi_crop}
\title{Slicing Aid for Hyper Inference (SAHI) for image}
\usage{
transform_sahi_crop(x, sahi_split)
}
\arguments{
\item{x}{Image input. Can be a \code{torch_tensor} of shape \verb{(C, H, W)}
or a \code{magick-image}.}

\item{sahi_split}{An object of class \code{sahi_split} created by
\code{prepare_sahi_split()}.}
}
\value{
An object of the same type as \code{x}:
\describe{
\item{torch_tensor in}{Stacked tensor of shape \verb{(N, C, H, W)}.}
\item{magick-image in}{Multi-frame magick image.}
}
}
\description{
Splits an image into overlapping crops using a precomputed
\code{sahi_split} object (created by \code{prepare_sahi_split()}).
Unlike other transforms, the split parameters are prepared once and
reused, keeping the transform type-invariant: the output has the same
type as the input (stacked tensor for tensor input, multi-frame magick
image for magick input).
}
\details{
Use as \code{transform} in image datasets.
}
\examples{
\dontrun{

img_url <- "https://raw.githubusercontent.com/obss/sahi/main/demo/demo_data/small-vehicles1.jpeg"

img <- base_loader(img_url) \%>\%
  transform_to_tensor()

sp <- prepare_sahi_split(img, size = c(512, 512), overlap_size_ratio = c(0.2, 0.2))

crops <- transform_sahi_crop(img, sp)

# crops is a stacked tensor of shape (N, C, H, W)
crops$shape

# Visualize the first 9 crops in a 3x3 grid
grid <- vision_make_grid(
  crops,
  scale = TRUE,
  num_rows = 3
)
tensor_image_browse(grid)
}

}
\seealso{
Other combining_transforms:
\code{\link[=transform_five_crop]{transform_five_crop()}},
\code{\link[=transform_random_apply]{transform_random_apply()}},
\code{\link[=transform_random_choice]{transform_random_choice()}},
\code{\link[=transform_random_order]{transform_random_order()}},
\code{\link[=transform_resized_crop]{transform_resized_crop()}},
\code{\link[=transform_ten_crop]{transform_ten_crop()}}
}
\concept{combining_transforms}
