% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-plankton.R
\name{whoi_small_coralnet_dataset}
\alias{whoi_small_coralnet_dataset}
\title{Coralnet Dataset}
\usage{
whoi_small_coralnet_dataset(
  split = "val",
  transform = NULL,
  target_transform = NULL,
  download = FALSE
)
}
\arguments{
\item{split}{One of \code{"train"}, \code{"val"}, or \code{"test"}. Default is \code{"val"}.}

\item{transform}{Optional. A function that takes an image and returns a transformed version (e.g., normalization, cropping).}

\item{target_transform}{Optional. A function that transforms the label.}

\item{download}{Logical. If TRUE, downloads the dataset to \verb{root/}. If the dataset is already present, download is skipped.}
}
\description{
Coralnet Dataset
}
